fix(studio): extract site-build design gates
