+++ Automattic/studio/bench/studio-agent-site-build.bench.mjs
@@ -6,6 +6,40 @@ import { spawn } from 'node:child_process';
 import { createRequire } from 'node:module';
 import { pathToFileURL } from 'node:url';
 
+import {
+  agentAuthoredBlockMetrics,
+  collectGeneratedThemeUxGates,
+  collectLatestImportReport,
+  collectThemeBlockDocuments,
+  nativeBlockQualityMetrics,
+  probeQuality,
+} from './lib/wordpress-quality.mjs';
+
+import { collectDesignFingerprint } from './lib/design-gates.mjs';
+
+export {
+  agentAuthoredBlockMetrics,
+  nativeBlockQualityMetrics,
+} from './lib/wordpress-quality.mjs';
+
+export { extractDesignPatternFingerprint } from './lib/design-gates.mjs';
+
+export {
+  hiddenEditorContentDiagnostics,
+  reportedFreeformBlockCount,
+  structuralSelectorDriftDiagnostics,
+} from './lib/design-gates.mjs';
+
+import { compareVisualFidelity, VISUAL_VIEWPORT } from './lib/visual-fidelity.mjs';
+export { compareVisualFidelity } from './lib/visual-fidelity.mjs';
+
+import {
+  compareSemanticFidelity as compareSemanticFidelityImpl,
+  semanticMismatchFailureDetails,
+  semanticTargetMetric,
+} from './lib/semantic-fidelity.mjs';
+export { semanticMismatchFailureDetails, semanticTargetMetric } from './lib/semantic-fidelity.mjs';
+
 const STUDIO_PATH = process.env.HOMEBOY_COMPONENT_PATH;
 const SHARED_STATE = process.env.HOMEBOY_BENCH_SHARED_STATE || os.tmpdir();
 const RESULT_PREFIX = 'EVAL_RUNNER_RESULT_FILE=';
@@ -16,39 +50,11 @@ const NAMESPACE_PORT_RANGE_COUNT = 100;
 const PROMPT_VARIANT_SETTING = 'studio_site_build_prompt_variant';
 const PROMPT_FILE_SETTING = 'studio_site_build_prompt_file';
 const BENCH_NAMESPACE_SETTING = 'studio_bench_namespace';
+const MODEL_SETTING = 'studio_agent_model';
 const DEFAULT_PROMPT_VARIANT = 'studio-code';
 const PROMPT_CATEGORY = 'site-build';
-const PROMPT_VARIANTS = {
-  'artist-music': 'plain-site/artist-music.md',
-  'course-education': 'plain-site/course-education.md',
-  'data-machine': 'plain-site/data-machine.md',
-  'documentation-knowledge-base': 'plain-site/documentation-knowledge-base.md',
-  'editorial-magazine': 'plain-site/editorial-magazine.md',
-  'event-conference': 'plain-site/event-conference.md',
-  homeboy: 'plain-site/homeboy.md',
-  intelligence: 'plain-site/intelligence.md',
-  'local-service-business': 'plain-site/local-service-business.md',
-  'membership-community': 'plain-site/membership-community.md',
-  nonprofit: 'plain-site/nonprofit.md',
-  'nonprofit-campaign': 'plain-site/nonprofit-campaign.md',
-  portfolio: 'plain-site/portfolio.md',
-  'radical-speed-month': 'plain-site/radical-speed-month.md',
-  restaurant: 'plain-site/restaurant.md',
-  saas: 'plain-site/saas.md',
-  'realistic-small-business': 'plain-site/realistic-small-business.md',
-  'studio-code': 'plain-site/studio-code.md',
-  'wp-coding-agents': 'plain-site/wp-coding-agents.md',
-  'wordpress-is-dead': 'plain-site/wordpress-is-dead.md',
-  'product-catalog': 'store/product-catalog.md',
-  'switchback-woocommerce-extra-hard': 'store/switchback-woocommerce-extra-hard.md',
-  'astro-docs-content-collection': 'static-markdown/astro-docs-content-collection.md',
-  'markdown-blog-launch-site': 'static-markdown/markdown-blog-launch-site.md',
-  'static-content-library': 'static-markdown/static-content-library.md',
-};
 const SYSTEM_PROMPT_FILES = ['apps/cli/ai/system-prompt.ts'];
 const requireFromBench = createRequire(import.meta.url);
-const VISUAL_VIEWPORT = { width: 1440, height: 1100 };
-const VISUAL_SCREENSHOT_DIAGNOSTIC_LIMIT = 5;
 
 if (!STUDIO_PATH) {
   throw new Error('HOMEBOY_COMPONENT_PATH is required');
@@ -78,6 +84,10 @@ function benchmarkNamespace() {
     process.env.studio_bench_namespace ||
     ''
   );
+}
+
+function evalModel() {
+  return setting(MODEL_SETTING) || process.env.STUDIO_EVAL_MODEL || '';
 }
 
 function namespacePortBase(namespace) {
@@ -296,30 +306,42 @@ function promptVariant() {
 }
 
 export async function availablePromptVariants() {
-  await validatePromptVariantCatalog();
-  return Object.keys(PROMPT_VARIANTS);
+  return validatePromptVariantCatalog();
 }
 
 export async function validatePromptVariantCatalog() {
   const promptsDir = new URL('./prompts/site-build/', import.meta.url);
   const files = await promptFiles(promptsDir);
-  const registeredEntries = Object.entries(PROMPT_VARIANTS);
-  const registered = registeredEntries.map(([variantName]) => variantName).sort();
-  const registeredFiles = registeredEntries.map(([, relativePath]) => relativePath).sort();
-  const missingFiles = registeredEntries
-    .filter(([, relativePath]) => !files.includes(relativePath))
-    .map(([variantName, relativePath]) => `${variantName} (${relativePath})`);
-  const unregisteredFiles = files.filter((relativePath) => !registeredFiles.includes(relativePath));
+  return Object.keys(promptVariantCatalog(files));
+}
 
-  if (missingFiles.length || unregisteredFiles.length) {
+export function promptVariantCatalog(files) {
+  const variants = {};
+  const duplicatePaths = new Map();
+
+  for (const relativePath of files) {
+    const variantName = path.posix.basename(relativePath, '.md');
+    if (variants[variantName]) {
+      if (!duplicatePaths.has(variantName)) {
+        duplicatePaths.set(variantName, [variants[variantName]]);
+      }
+      duplicatePaths.get(variantName).push(relativePath);
+      continue;
+    }
+
+    variants[variantName] = relativePath;
+  }
+
+  if (duplicatePaths.size) {
+    const duplicates = [...duplicatePaths.entries()]
+      .map(([variantName, paths]) => `${variantName} (${paths.join(', ')})`)
+      .join('; ');
     throw new Error(
-      `Studio site-build prompt catalog mismatch. Missing files: ${
-        missingFiles.join(', ') || 'none'
-      }. Unregistered files: ${unregisteredFiles.join(', ') || 'none'}.`
+      `Studio site-build prompt catalog has duplicate basename-derived variant IDs. Rename prompt files so each basename is unique. Duplicates: ${duplicates}.`
     );
   }
 
-  return registered;
+  return variants;
 }
 
 async function promptFiles(directoryUrl, prefix = '') {
@@ -348,7 +370,8 @@ async function promptTemplatePath() {
   if (!/^[a-z0-9][a-z0-9-]*$/.test(variantName)) {
     throw new Error(`Invalid ${PROMPT_VARIANT_SETTING}: ${variantName}`);
   }
-  const relativePromptPath = PROMPT_VARIANTS[variantName];
+  const promptsDir = new URL('./prompts/site-build/', import.meta.url);
+  const relativePromptPath = promptVariantCatalog(await promptFiles(promptsDir))[variantName];
   if (!relativePromptPath) {
     const variants = await availablePromptVariants();
     throw new Error(`Unknown ${PROMPT_VARIANT_SETTING}: ${variantName}. Available variants: ${variants.join(', ')}.`);
@@ -420,101 +443,6 @@ async function createFreshSite(sitePath) {
   ]);
 }
 
-const QUALITY_PROBE = String.raw`
-function bench_count_blocks( $blocks, &$counts ) {
-    foreach ( $blocks as $block ) {
-        $name = isset( $block['blockName'] ) ? (string) $block['blockName'] : '';
-        if ( '' !== $name ) {
-            $counts['total_blocks']++;
-            if ( 'core/html' === $name ) {
-                $counts['core_html_blocks']++;
-            }
-        }
-        if ( ! empty( $block['innerBlocks'] ) ) {
-            bench_count_blocks( $block['innerBlocks'], $counts );
-        }
-    }
-}
-
-$counts = array(
-    'posts_seen' => 0,
-    'posts_with_blocks' => 0,
-    'pages_seen' => 0,
-    'templates_seen' => 0,
-    'template_parts_seen' => 0,
-    'target_pages_seen' => 0,
-    'target_posts_with_blocks' => 0,
-    'target_raw_html_unconverted' => 0,
-    'target_total_blocks' => 0,
-    'target_core_html_blocks' => 0,
-    'target_serialized_block_comments' => 0,
-    'total_blocks' => 0,
-    'core_html_blocks' => 0,
-    'serialized_block_comments' => 0,
-    'bfb_fallback_count' => (int) get_option( 'studio_bfb_unsupported_fallback_count', 0 ),
-);
-
-$front_page_id = (int) get_option( 'page_on_front', 0 );
-
-$posts = get_posts( array(
-    'post_type' => array( 'page', 'wp_template', 'wp_template_part' ),
-    'post_status' => 'any',
-    'numberposts' => -1,
-) );
-
-foreach ( $posts as $post ) {
-    $content = (string) $post->post_content;
-    if ( '' === trim( $content ) ) {
-        continue;
-    }
-    $counts['posts_seen']++;
-    if ( 'page' === $post->post_type ) {
-        $counts['pages_seen']++;
-    }
-    if ( 'wp_template' === $post->post_type ) {
-        $counts['templates_seen']++;
-    }
-    if ( 'wp_template_part' === $post->post_type ) {
-        $counts['template_parts_seen']++;
-    }
-    $counts['serialized_block_comments'] += substr_count( $content, '<!-- wp:' );
-    if ( false !== strpos( $content, '<!-- wp:' ) ) {
-        $counts['posts_with_blocks']++;
-    }
-    $before_total     = $counts['total_blocks'];
-    $before_core_html = $counts['core_html_blocks'];
-    bench_count_blocks( parse_blocks( $content ), $counts );
-
-    $is_target_page = 'page' === $post->post_type && ( (int) $post->ID === $front_page_id || 'Studio Code' === $post->post_title );
-    if ( $is_target_page ) {
-        $counts['target_pages_seen']++;
-        $counts['target_serialized_block_comments'] += substr_count( $content, '<!-- wp:' );
-        if ( false !== strpos( $content, '<!-- wp:' ) ) {
-            $counts['target_posts_with_blocks']++;
-        }
-        if ( false === strpos( $content, '<!-- wp:' ) && preg_match( '/<\/?[a-z][\s>]/i', $content ) ) {
-            $counts['target_raw_html_unconverted']++;
-        }
-        $counts['target_total_blocks'] += $counts['total_blocks'] - $before_total;
-        $counts['target_core_html_blocks'] += $counts['core_html_blocks'] - $before_core_html;
-    }
-}
-
-$counts['core_html_without_bfb_fallback'] = max( 0, $counts['core_html_blocks'] - $counts['bfb_fallback_count'] );
-$counts['target_core_html_without_bfb_fallback'] = max( 0, $counts['target_core_html_blocks'] - $counts['bfb_fallback_count'] );
-
-echo wp_json_encode( $counts, JSON_PRETTY_PRINT ) . PHP_EOL;
-`;
-
-async function probeQuality(sitePath) {
-  const { stdout } = await runCli(['wp', '--path', sitePath, '--php-version', '8.3', 'eval', QUALITY_PROBE]);
-  const jsonStart = stdout.indexOf('{');
-  if (jsonStart === -1) {
-    throw new Error(`quality probe did not emit JSON: ${stdout.slice(0, 1000)}`);
-  }
-  return JSON.parse(stdout.slice(jsonStart));
-}
-
 async function siteStatus(sitePath) {
   const { stdout } = await runCli(['site', 'status', '--path', sitePath, '--format', 'json']);
   const jsonStart = stdout.indexOf('{');
@@ -522,93 +450,6 @@ async function siteStatus(sitePath) {
     throw new Error(`site status did not emit JSON: ${stdout.slice(0, 1000)}`);
   }
   return JSON.parse(stdout.slice(jsonStart));
-}
-
-async function collectThemeBlockDocuments(sitePath) {
-  const { themeRoot } = await collectLatestGeneratedTheme(sitePath);
-  if (!themeRoot) {
-    return [];
-  }
-
-  const documents = [];
-
-  async function collectDir(relativeDir, extension) {
-    const dir = path.join(themeRoot, relativeDir);
-    let entries = [];
-    try {
-      entries = await readdir(dir, { withFileTypes: true });
-    } catch {
-      return;
-    }
-
-    for (const entry of entries) {
-      if (!entry.isFile() || !entry.name.endsWith(extension)) {
-        continue;
-      }
-      const file = path.join(dir, entry.name);
-      let content = await readFile(file, 'utf8');
-      if (extension === '.php') {
-        content = content.replace(/^\s*<\?php[\s\S]*?\?>\s*/, '');
-      }
-      if (content.includes('<!-- wp:')) {
-        documents.push({
-          source: path.relative(themeRoot, file),
-          content,
-        });
-      }
-    }
-  }
-
-  await collectDir('templates', '.html');
-  await collectDir('parts', '.html');
-  await collectDir('patterns', '.php');
-
-  return documents;
-}
-
-async function collectLatestGeneratedTheme(sitePath) {
-  const themesRoot = path.join(sitePath, 'wp-content/themes');
-  let themeEntries = [];
-  try {
-    themeEntries = await readdir(themesRoot, { withFileTypes: true });
-  } catch {
-    return { themeRoot: '', themeSlug: '', reportPath: '', error: 'Themes directory not found.' };
-  }
-
-  const generatedThemes = [];
-  for (const entry of themeEntries) {
-    if (!entry.isDirectory()) {
-      continue;
-    }
-    const themeRoot = path.join(themesRoot, entry.name);
-    const reportPath = path.join(themeRoot, 'import-report.json');
-    try {
-      const reportStat = await stat(reportPath);
-      generatedThemes.push({ themeRoot, themeSlug: entry.name, reportPath, mtimeMs: reportStat.mtimeMs });
-    } catch {
-      // Ignore non-importer themes.
-    }
-  }
-
-  generatedThemes.sort((a, b) => b.mtimeMs - a.mtimeMs);
-  return generatedThemes[0] || { themeRoot: '', themeSlug: '', reportPath: '', error: 'No Static Site Importer report found.' };
-}
-
-async function collectLatestImportReport(sitePath) {
-  const { reportPath, error } = await collectLatestGeneratedTheme(sitePath);
-  if (!reportPath) {
-    return { report: null, reportPath: '', error };
-  }
-
-  try {
-    return { report: JSON.parse(await readFile(reportPath, 'utf8')), reportPath, error: '' };
-  } catch (error) {
-    return {
-      report: null,
-      reportPath,
-      error: `Failed to read Static Site Importer report: ${error instanceof Error ? error.message : String(error)}`,
-    };
-  }
 }
 
 function asArray(value) {
@@ -634,12 +475,6 @@ function comparisonTargets(importReport) {
   );
 }
 
-function semanticComparisonTargets(importReport) {
-  const semanticTargets = asArray(importReport?.report?.semantic_fidelity?.comparison_targets).filter(
-    (target) => target && typeof target === 'object'
-  );
-  return semanticTargets.length ? semanticTargets : comparisonTargets(importReport);
-}
 
 export function resolveSourceStaticFile(sourceFile, reportPath, sitePath) {
   if (!sourceFile) {
@@ -715,2064 +550,11 @@ async function restoreMissingSourceStaticFiles(importReport, sitePath, result) {
   }
 }
 
-function visualProbeGroups(target) {
-  const hooks = target?.comparison_hooks || {};
-  const layoutProbes = hooks.layout_probes && typeof hooks.layout_probes === 'object' ? hooks.layout_probes : {};
-  const groups = [];
-  const seen = new Set();
-
-  function add(name, selectors) {
-    const normalizedSelectors = stringArray(selectors);
-    if (!normalizedSelectors.length || seen.has(name)) {
-      return;
-    }
-    seen.add(name);
-    groups.push({ name, selectors: normalizedSelectors });
-  }
-
-  for (const [name, probe] of Object.entries(layoutProbes)) {
-    add(name, probe?.selectors);
-  }
-
-  add('hero_probe', hooks.hero);
-  add('visible_chrome', hooks.visible_chrome);
-  add('footer_chrome', ['footer', '.site-footer', '[class*=footer]']);
-
-  return groups;
-}
-
-async function loadVisualSurface(page, url) {
-  await page.goto(url, { waitUntil: 'domcontentloaded', timeout: 45_000 });
-  await page.waitForLoadState('networkidle', { timeout: 5_000 }).catch(() => {});
-  await page.waitForTimeout(300);
-}
-
-async function evaluateVisualSurface(page, groups) {
-  const evaluatedGroups = [];
-
-  for (const group of groups) {
-    const selectors = [];
-
-    for (const selector of group.selectors) {
-      try {
-        const matches = await page.$$eval(selector, (elements) =>
-          elements.map((element) => {
-            const rect = element.getBoundingClientRect();
-            const style = window.getComputedStyle(element);
-            const visible =
-              rect.width > 0 &&
-              rect.height > 0 &&
-              style.display !== 'none' &&
-              style.visibility !== 'hidden' &&
-              Number(style.opacity || '1') > 0;
-
-            return {
-              visible,
-              boundingBox: {
-                x: Math.round(rect.x),
-                y: Math.round(rect.y),
-                width: Math.round(rect.width),
-                height: Math.round(rect.height),
-              },
-              text: String(element.textContent || '').trim().replace(/\s+/g, ' ').slice(0, 160),
-            };
-          })
-        );
-        selectors.push({
-          selector,
-          count: matches.length,
-          visible_count: matches.filter((match) => match.visible).length,
-          nonzero_bounding_box_count: matches.filter(
-            (match) => match.boundingBox.width > 0 && match.boundingBox.height > 0
-          ).length,
-          first_match: matches[0] || null,
-        });
-      } catch (error) {
-        selectors.push({
-          selector,
-          count: 0,
-          visible_count: 0,
-          nonzero_bounding_box_count: 0,
-          error: error instanceof Error ? error.message : String(error),
-          first_match: null,
-        });
-      }
-    }
-
-    evaluatedGroups.push({
-      name: group.name,
-      selectors,
-      selector_count: selectors.length,
-      missing_selector_count: selectors.filter((item) => item.count === 0).length,
-      errored_selector_count: selectors.filter((item) => item.error).length,
-      matched_selector_count: selectors.filter((item) => item.count > 0).length,
-      visible_selector_count: selectors.filter((item) => item.visible_count > 0).length,
-      nonzero_bounding_box_selector_count: selectors.filter((item) => item.nonzero_bounding_box_count > 0).length,
-    });
-  }
-
-  return evaluatedGroups;
-}
-
-function visualSurfaceTotals(groups) {
-  return groups.reduce(
-    (totals, group) => {
-      totals.selector_count += group.selector_count;
-      totals.missing_selector_count += group.missing_selector_count;
-      totals.errored_selector_count += group.errored_selector_count;
-      totals.matched_selector_count += group.matched_selector_count;
-      totals.visible_selector_count += group.visible_selector_count;
-      totals.nonzero_bounding_box_selector_count += group.nonzero_bounding_box_selector_count;
-      return totals;
-    },
-    {
-      selector_count: 0,
-      missing_selector_count: 0,
-      errored_selector_count: 0,
-      matched_selector_count: 0,
-      visible_selector_count: 0,
-      nonzero_bounding_box_selector_count: 0,
-    }
-  );
-}
-
-function visualSelectorSummary(selector) {
-  const firstMatch = selector?.first_match || null;
-  return {
-    count: Number(selector?.count || 0),
-    visible_count: Number(selector?.visible_count || 0),
-    nonzero_bounding_box_count: Number(selector?.nonzero_bounding_box_count || 0),
-    first_bounding_box: firstMatch?.boundingBox || null,
-    first_visible: firstMatch?.visible === true,
-    first_visible_text: firstMatch?.text || '',
-    error: selector?.error || '',
-  };
-}
-
-function visualMismatchReason(sourceSelector, frontendSelector) {
-  if (sourceSelector?.error || frontendSelector?.error) {
-    return 'selector_error';
-  }
-  if (sourceSelector.visible_count === 0 && frontendSelector.visible_count === 0) {
-    return 'missing_on_both_surfaces';
-  }
-  if (sourceSelector.count === 0 && frontendSelector.count === 0) {
-    return 'missing_on_both_surfaces';
-  }
-  if (sourceSelector.count === 0) {
-    return 'missing_from_source_static';
-  }
-  if (frontendSelector.count === 0) {
-    return 'missing_from_wordpress_frontend';
-  }
-
-  const sourceVisible = sourceSelector.visible_count > 0;
-  const frontendVisible = frontendSelector.visible_count > 0;
-  if (sourceVisible !== frontendVisible) {
-    return sourceVisible ? 'hidden_on_wordpress_frontend' : 'hidden_on_source_static';
-  }
-
-  const sourceNonzero = sourceSelector.nonzero_bounding_box_count > 0;
-  const frontendNonzero = frontendSelector.nonzero_bounding_box_count > 0;
-  if (sourceNonzero !== frontendNonzero) {
-    return sourceNonzero ? 'zero_sized_on_wordpress_frontend' : 'zero_sized_on_source_static';
-  }
-
-  return '';
-}
-
-function visualMismatchSeverity(reason) {
-  const severities = {
-    selector_error: 100,
-    missing_from_wordpress_frontend: 90,
-    missing_from_source_static: 80,
-    hidden_on_wordpress_frontend: 60,
-    hidden_on_source_static: 50,
-    zero_sized_on_wordpress_frontend: 40,
-    zero_sized_on_source_static: 30,
-  };
-  return severities[reason] || 0;
-}
-
-function visualGroupMismatchSummary(groupName, mismatches) {
-  const reasons = {};
-  for (const mismatch of mismatches) {
-    reasons[mismatch.reason] = (reasons[mismatch.reason] || 0) + 1;
-  }
-  return {
-    group: groupName,
-    mismatch_count: mismatches.length,
-    reasons,
-    top_selectors: mismatches.slice(0, 5).map((mismatch) => ({
-      selector: mismatch.selector,
-      reason: mismatch.reason,
-      source_count: mismatch.source.count,
-      frontend_count: mismatch.frontend.count,
-      source_visible_count: mismatch.source.visible_count,
-      frontend_visible_count: mismatch.frontend.visible_count,
-    })),
-  };
-}
-
-function visualSelectorComparisonDetail(sourceGroup, sourceSelector, frontendSelector, reason) {
-  return {
-    group: sourceGroup.name,
-    selector: sourceSelector.selector,
-    reason,
-    severity: visualMismatchSeverity(reason),
-    source: visualSelectorSummary(sourceSelector),
-    frontend: visualSelectorSummary(frontendSelector),
-    screenshots: {},
-  };
-}
-
-function visualSelectorComparisonDetails(result) {
-  const sourceGroups = result.surfaces.source_static?.probes || [];
-  const frontendGroups = result.surfaces.wordpress_frontend?.probes || [];
-  const frontendGroupsByName = new Map(frontendGroups.map((group) => [group.name, group]));
-  const mismatches = [];
-  const optionalProbeAbsences = [];
-
-  for (const sourceGroup of sourceGroups) {
-    const frontendGroup = frontendGroupsByName.get(sourceGroup.name);
-    if (!frontendGroup) {
-      continue;
-    }
-
-    const frontendSelectors = new Map(frontendGroup.selectors.map((selector) => [selector.selector, selector]));
-    for (const sourceSelector of sourceGroup.selectors) {
-      const frontendSelector = frontendSelectors.get(sourceSelector.selector);
-      if (!frontendSelector) {
-        continue;
-      }
-
-      const reason = visualMismatchReason(sourceSelector, frontendSelector);
-      if (!reason) {
-        continue;
-      }
-
-      const detail = visualSelectorComparisonDetail(sourceGroup, sourceSelector, frontendSelector, reason);
-      if (reason === 'missing_on_both_surfaces') {
-        optionalProbeAbsences.push(detail);
-        continue;
-      }
-
-      mismatches.push(detail);
-    }
-  }
-
-  mismatches.sort(
-    (a, b) => b.severity - a.severity || a.group.localeCompare(b.group) || a.selector.localeCompare(b.selector)
-  );
-  optionalProbeAbsences.sort((a, b) => a.group.localeCompare(b.group) || a.selector.localeCompare(b.selector));
-  return { mismatches, optionalProbeAbsences };
-}
-
-async function captureSelectorScreenshot(page, selector, screenshotPath) {
-  try {
-    const locator = page.locator(selector).first();
-    if ((await locator.count()) === 0) {
-      return '';
-    }
-    await locator.screenshot({ path: screenshotPath, timeout: 5_000 });
-    return screenshotPath;
-  } catch {
-    return '';
-  }
-}
-
-async function captureVisualMismatchScreenshots(browser, result, mismatches, visualDir, targetSlug) {
-  if (!mismatches.length) {
-    return;
-  }
-
-  const screenshotsDir = path.join(visualDir, 'mismatch-screenshots');
-  const urls = {
-    source_static: result.surfaces.source_static?.url || '',
-    wordpress_frontend: result.surfaces.wordpress_frontend?.url || '',
-  };
-  const pages = {};
-
-  try {
-    await mkdir(screenshotsDir, { recursive: true });
-    for (const [surface, url] of Object.entries(urls)) {
-      if (!url) {
-        continue;
-      }
-      const page = await browser.newPage({ ignoreHTTPSErrors: true, viewport: VISUAL_VIEWPORT });
-      await loadVisualSurface(page, url);
-      pages[surface] = page;
-    }
-
-    for (const [index, mismatch] of mismatches.slice(0, VISUAL_SCREENSHOT_DIAGNOSTIC_LIMIT).entries()) {
-      const screenshotSlug = safeSlug(
-        `${targetSlug}-${index + 1}-${mismatch.group}-${mismatch.selector}`,
-        `mismatch-${index + 1}`
-      );
-      for (const [surface, page] of Object.entries(pages)) {
-        const screenshotPath = path.join(screenshotsDir, `${screenshotSlug}-${surface}.png`);
-        const capturedPath = await captureSelectorScreenshot(page, mismatch.selector, screenshotPath);
-        if (capturedPath) {
-          mismatch.screenshots[surface] = capturedPath;
-        }
-      }
-    }
-  } catch (error) {
-    result.diagnostic_warnings = [
-      ...(result.diagnostic_warnings || []),
-      `mismatch screenshots: ${error instanceof Error ? error.message : String(error)}`,
-    ];
-  } finally {
-    await Promise.all(Object.values(pages).map((page) => page.close().catch(() => {})));
-  }
-}
-
-function buildVisualDiagnostics(results, artifactPath) {
-  const targetSummaries = [];
-  const allMismatches = [];
-  const allOptionalProbeAbsences = [];
-
-  for (const result of results) {
-    const mismatches = asArray(result.diagnostics?.mismatches);
-    const optionalProbeAbsences = asArray(result.diagnostics?.optional_probe_absences);
-    const byGroup = new Map();
-
-    for (const mismatch of mismatches) {
-      if (!byGroup.has(mismatch.group)) {
-        byGroup.set(mismatch.group, []);
-      }
-      byGroup.get(mismatch.group).push(mismatch);
-      allMismatches.push({
-        target: result.source_filename || String(result.wordpress_page_id || ''),
-        ...mismatch,
-      });
-    }
-
-    for (const absence of optionalProbeAbsences) {
-      allOptionalProbeAbsences.push({
-        target: result.source_filename || String(result.wordpress_page_id || ''),
-        ...absence,
-      });
-    }
-
-    targetSummaries.push({
-      source_filename: result.source_filename || '',
-      wordpress_page_id: result.wordpress_page_id || null,
-      mismatch_count: mismatches.length,
-      optional_probe_absent_count: optionalProbeAbsences.length,
-      top_failing_groups: [...byGroup.entries()]
-        .map(([groupName, groupMismatches]) => visualGroupMismatchSummary(groupName, groupMismatches))
-        .sort((a, b) => b.mismatch_count - a.mismatch_count || a.group.localeCompare(b.group))
-        .slice(0, 5),
-      top_failing_selectors: mismatches.slice(0, 10).map((mismatch) => ({
-        group: mismatch.group,
-        selector: mismatch.selector,
-        reason: mismatch.reason,
-        source_count: mismatch.source.count,
-        frontend_count: mismatch.frontend.count,
-        source_first_bounding_box: mismatch.source.first_bounding_box,
-        frontend_first_bounding_box: mismatch.frontend.first_bounding_box,
-        source_first_visible_text: mismatch.source.first_visible_text,
-        frontend_first_visible_text: mismatch.frontend.first_visible_text,
-        screenshots: mismatch.screenshots,
-      })),
-    });
-  }
-
-  const topFailingGroups = new Map();
-  for (const mismatch of allMismatches) {
-    const key = `${mismatch.target || 'target'}:${mismatch.group}`;
-    if (!topFailingGroups.has(key)) {
-      topFailingGroups.set(key, []);
-    }
-    topFailingGroups.get(key).push(mismatch);
-  }
-
-  return {
-    artifact: artifactPath,
-    mismatch_count: allMismatches.length,
-    optional_probe_absent_count: allOptionalProbeAbsences.length,
-    top_failing_groups: [...topFailingGroups.entries()]
-      .map(([, mismatches]) => ({
-        target: mismatches[0]?.target || '',
-        ...visualGroupMismatchSummary(mismatches[0]?.group || '', mismatches),
-      }))
-      .sort((a, b) => b.mismatch_count - a.mismatch_count || a.group.localeCompare(b.group))
-      .slice(0, 10),
-    targets: targetSummaries,
-    mismatches: allMismatches,
-    optional_probe_absences: allOptionalProbeAbsences,
-  };
-}
-
-function visualParity(sourceGroups, frontendGroups) {
-  const frontendByName = new Map(frontendGroups.map((group) => [group.name, group]));
-  const groupComparisons = [];
-  let missingSelectorCount = 0;
-  let visibilityMismatchCount = 0;
-  let nonzeroBoundingBoxMismatchCount = 0;
-  let simpleProbeParityMismatchCount = 0;
-  const simpleProbeFamilies = {
-    nav: new Set(['nav_chrome']),
-    footer: new Set(['footer_chrome']),
-    hero: new Set(['hero_region', 'hero_probe']),
-  };
-  const simpleProbeNames = new Set(Object.values(simpleProbeFamilies).flatMap((names) => [...names]));
-  const simpleProbeMismatches = { nav: 0, footer: 0, hero: 0 };
-
-  for (const sourceGroup of sourceGroups) {
-    const frontendGroup = frontendByName.get(sourceGroup.name);
-    if (!frontendGroup) {
-      continue;
-    }
-
-    const frontendSelectors = new Map(frontendGroup.selectors.map((selector) => [selector.selector, selector]));
-    for (const sourceSelector of sourceGroup.selectors) {
-      const frontendSelector = frontendSelectors.get(sourceSelector.selector);
-      if (!frontendSelector) {
-        continue;
-      }
-
-      const sourceVisible = sourceSelector.visible_count > 0;
-      const frontendVisible = frontendSelector.visible_count > 0;
-      const sourceNonzero = sourceSelector.nonzero_bounding_box_count > 0;
-      const frontendNonzero = frontendSelector.nonzero_bounding_box_count > 0;
-
-      if (sourceSelector.count === 0 || frontendSelector.count === 0) {
-        missingSelectorCount++;
-      }
-      if (sourceVisible !== frontendVisible) {
-        visibilityMismatchCount++;
-      }
-      if (sourceNonzero !== frontendNonzero) {
-        nonzeroBoundingBoxMismatchCount++;
-      }
-    }
-
-    const sourceGroupVisible = sourceGroup.visible_selector_count > 0;
-    const frontendGroupVisible = frontendGroup.visible_selector_count > 0;
-    const simpleProbeMismatch = simpleProbeNames.has(sourceGroup.name) && sourceGroupVisible !== frontendGroupVisible;
-    if (simpleProbeMismatch) {
-      simpleProbeParityMismatchCount++;
-      for (const [family, names] of Object.entries(simpleProbeFamilies)) {
-        if (names.has(sourceGroup.name)) {
-          simpleProbeMismatches[family]++;
-        }
-      }
-    }
-
-    groupComparisons.push({
-      name: sourceGroup.name,
-      source_visible: sourceGroupVisible,
-      frontend_visible: frontendGroupVisible,
-      source_nonzero_bounding_box: sourceGroup.nonzero_bounding_box_selector_count > 0,
-      frontend_nonzero_bounding_box: frontendGroup.nonzero_bounding_box_selector_count > 0,
-      simple_probe_parity: simpleProbeNames.has(sourceGroup.name) ? !simpleProbeMismatch : null,
-    });
-  }
-
-  return {
-    missing_selector_count: missingSelectorCount,
-    visibility_mismatch_count: visibilityMismatchCount,
-    nonzero_bounding_box_mismatch_count: nonzeroBoundingBoxMismatchCount,
-    simple_probe_parity_mismatch_count: simpleProbeParityMismatchCount,
-    simple_probe_mismatches: simpleProbeMismatches,
-    groups: groupComparisons,
-  };
-}
-
-async function emptyVisualComparison(artifactDir, error = '') {
-  const visualDir = path.join(artifactDir, 'visual-comparisons');
-  await mkdir(visualDir, { recursive: true });
-  const diagnosticsPath = path.join(visualDir, 'visual-comparison-skipped.json');
-  const diagnostics = {
-    ...buildVisualDiagnostics([], diagnosticsPath),
-    skipped: true,
-    reason: error,
-  };
-  await writeFile(diagnosticsPath, JSON.stringify(diagnostics, null, 2));
-
-  return {
-    target_count: 0,
-    checked_target_count: 0,
-    error_count: error ? 1 : 0,
-    missing_selector_count: 0,
-    visibility_mismatch_count: 0,
-    nonzero_bounding_box_count: 0,
-    nonzero_bounding_box_mismatch_count: 0,
-    simple_probe_parity_mismatch_count: 0,
-    nav_probe_parity_mismatch_count: 0,
-    footer_probe_parity_mismatch_count: 0,
-    hero_probe_parity_mismatch_count: 0,
-    surfaces: ['source_static', 'wordpress_frontend'],
-    editor_surface_ready: true,
-    artifact_dir: visualDir,
-    diagnostics_artifact: diagnosticsPath,
-    error,
-    results: [],
-    diagnostics,
-  };
-}
-
-export async function compareVisualFidelity(importReport, artifactDir, sitePath) {
-  const targets = comparisonTargets(importReport);
-  if (!targets.length) {
-    return emptyVisualComparison(artifactDir, importReport?.error || 'No visual fidelity comparison targets found.');
-  }
-
-  const playwrightPackage = path.join(STUDIO_PATH, 'node_modules/@playwright/test');
-  const { chromium } = requireFromBench(playwrightPackage);
-  const visualDir = path.join(artifactDir, 'visual-comparisons');
-  await mkdir(visualDir, { recursive: true });
-
-  let browser;
-  const results = [];
-
-  try {
-    browser = await chromium.launch({ args: ['--ignore-certificate-errors'] });
-
-    for (const [index, target] of targets.entries()) {
-      const targetSlug = safeSlug(target.source_filename || target.wordpress_page_id, `target-${index + 1}`);
-      const groups = visualProbeGroups(target);
-      const result = {
-        source_filename: target.source_filename || '',
-        wordpress_page_id: target.wordpress_page_id || null,
-        generated_template: target.generated_template || '',
-        generated_pattern: target.generated_pattern || '',
-        comparison_hooks: target.comparison_hooks || {},
-        source_probe_counts: target.source_probe_counts || {},
-        generated_probe_counts: target.generated_probe_counts || {},
-        surfaces: {},
-        parity: null,
-        errors: [],
-      };
-
-      for (const surface of ['source_static', 'wordpress_frontend']) {
-        const url = surfaceUrl(target, surface, importReport.reportPath, sitePath);
-        const screenshotPath = path.join(visualDir, `${targetSlug}-${surface}.png`);
-        const page = await browser.newPage({ ignoreHTTPSErrors: true, viewport: VISUAL_VIEWPORT });
-
-        try {
-          if (!url) {
-            throw new Error(`Missing ${surface} render URL.`);
-          }
-          await loadVisualSurface(page, url);
-          const probeGroups = await evaluateVisualSurface(page, groups);
-          await page.screenshot({ path: screenshotPath, fullPage: true });
-          result.surfaces[surface] = {
-            url,
-            screenshot: screenshotPath,
-            probes: probeGroups,
-            totals: visualSurfaceTotals(probeGroups),
-          };
-        } catch (error) {
-          const message = error instanceof Error ? error.message : String(error);
-          result.errors.push(`${surface}: ${message}`);
-          result.surfaces[surface] = { url, screenshot: '', probes: [], totals: visualSurfaceTotals([]), error: message };
-        } finally {
-          await page.close();
-        }
-      }
-
-      result.parity = visualParity(
-        result.surfaces.source_static?.probes || [],
-        result.surfaces.wordpress_frontend?.probes || []
-      );
-      const { mismatches, optionalProbeAbsences } = visualSelectorComparisonDetails(result);
-      result.diagnostics = {
-        mismatch_count: mismatches.length,
-        optional_probe_absent_count: optionalProbeAbsences.length,
-        top_failing_groups: [],
-        mismatches,
-        optional_probe_absences: optionalProbeAbsences,
-      };
-      await captureVisualMismatchScreenshots(browser, result, mismatches, visualDir, targetSlug);
-      results.push(result);
-    }
-  } catch (error) {
-    return emptyVisualComparison(artifactDir, `Visual comparison failed: ${error instanceof Error ? error.message : String(error)}`);
-  } finally {
-    if (browser) {
-      await browser.close();
-    }
-  }
-
-  const totals = results.reduce(
-    (summary, result) => {
-      const sourceTotals = result.surfaces.source_static?.totals || visualSurfaceTotals([]);
-      const frontendTotals = result.surfaces.wordpress_frontend?.totals || visualSurfaceTotals([]);
-      const parity = result.parity || visualParity([], []);
-      summary.error_count += result.errors.length;
-      summary.checked_target_count += result.errors.length === 0 ? 1 : 0;
-      summary.missing_selector_count += sourceTotals.missing_selector_count + frontendTotals.missing_selector_count;
-      summary.visibility_mismatch_count += parity.visibility_mismatch_count;
-      summary.nonzero_bounding_box_count +=
-        sourceTotals.nonzero_bounding_box_selector_count + frontendTotals.nonzero_bounding_box_selector_count;
-      summary.nonzero_bounding_box_mismatch_count += parity.nonzero_bounding_box_mismatch_count;
-      summary.simple_probe_parity_mismatch_count += parity.simple_probe_parity_mismatch_count;
-      summary.nav_probe_parity_mismatch_count += parity.simple_probe_mismatches?.nav || 0;
-      summary.footer_probe_parity_mismatch_count += parity.simple_probe_mismatches?.footer || 0;
-      summary.hero_probe_parity_mismatch_count += parity.simple_probe_mismatches?.hero || 0;
-      return summary;
-    },
-    {
-      target_count: targets.length,
-      checked_target_count: 0,
-      error_count: 0,
-      missing_selector_count: 0,
-      visibility_mismatch_count: 0,
-      nonzero_bounding_box_count: 0,
-      nonzero_bounding_box_mismatch_count: 0,
-      simple_probe_parity_mismatch_count: 0,
-      nav_probe_parity_mismatch_count: 0,
-      footer_probe_parity_mismatch_count: 0,
-      hero_probe_parity_mismatch_count: 0,
-    }
-  );
-  const diagnosticsPath = path.join(visualDir, 'visual-comparison-mismatches.json');
-  const diagnostics = buildVisualDiagnostics(results, diagnosticsPath);
-  await writeFile(diagnosticsPath, JSON.stringify(diagnostics, null, 2));
-
-  return {
-    ...totals,
-    surfaces: ['source_static', 'wordpress_frontend'],
-    editor_surface_ready: true,
-    artifact_dir: visualDir,
-    diagnostics_artifact: diagnosticsPath,
-    diagnostics,
-    results,
-  };
-}
-
-function emptySemanticComparison(error = '', artifactPath = '', diagnostics = null) {
-  return {
-    target_count: 0,
-    checked_target_count: 0,
-    error_count: error ? 1 : 0,
-    mismatch_count: 0,
-    role_mismatch_count: 0,
-    class_owner_changed_count: 0,
-    interaction_group_split_count: 0,
-    interaction_group_merged_count: 0,
-    link_text_delta_count: 0,
-    region_link_count_delta: 0,
-    clickable_area_delta_ratio: 0,
-    landmark_mismatch_count: 0,
-    repeated_count_delta_count: 0,
-    brand_logo_missing_count: 0,
-    error,
-    results: [],
-    ...(artifactPath ? { artifact_dir: path.dirname(artifactPath), artifact: artifactPath } : {}),
-    ...(diagnostics ? { diagnostics } : {}),
-  };
-}
-
-function semanticTargetSelectorGroups(target) {
-  const hooks = target?.comparison_hooks || {};
-  const layoutProbes = hooks.layout_probes && typeof hooks.layout_probes === 'object' ? hooks.layout_probes : {};
-  const groups = [];
-  const seen = new Set();
-
-  function add(name, selectors) {
-    const normalizedSelectors = stringArray(selectors);
-    if (!normalizedSelectors.length || seen.has(name)) {
-      return;
-    }
-    seen.add(name);
-    groups.push({ name, selectors: normalizedSelectors });
-  }
-
-  for (const [name, probe] of Object.entries(layoutProbes)) {
-    add(name, probe?.selectors);
-  }
-
-  add('hero', hooks.hero);
-  add('visible_chrome', hooks.visible_chrome);
-  add('footer_chrome', ['footer', '.site-footer', '[class*=footer]']);
-  add('brand_hooks', ['[class*=brand]', '[class*=logo]', '[class*=wordmark]']);
-  add('interaction_hooks', ['a', 'button', '[role=button]', '[role=link]']);
-
-  return groups;
-}
-
-function semanticSurfaceTotals(fingerprint) {
-  const regions = fingerprint?.regions || {};
-  return Object.values(regions).reduce(
-    (totals, region) => {
-      totals.region_link_count += Number(region?.link_count || 0);
-      totals.clickable_area += Number(region?.clickable_area || 0);
-      return totals;
-    },
-    { region_link_count: 0, clickable_area: 0 }
-  );
-}
-
-function semanticFingerprintExtractor(groups) {
-  const meaningfulHookPattern = /(brand|logo|wordmark|nav|menu|footer|header|hero|card|panel|cta|button|price|plan|feature|testimonial|avatar|badge|label|eyebrow|status|icon)/i;
-  const landmarkSelectors = {
-    header: 'header,[role=banner]',
-    nav: 'nav,[role=navigation]',
-    main: 'main,[role=main]',
-    footer: 'footer,[role=contentinfo]',
-    section: 'section',
-    aside: 'aside,[role=complementary]',
-  };
-
-  function normalizeText(value, limit = 180) {
-    return String(value || '').trim().replace(/\s+/g, ' ').slice(0, limit);
-  }
-
-  function visible(element) {
-    const rect = element.getBoundingClientRect();
-    const style = window.getComputedStyle(element);
-    return (
-      rect.width > 0 &&
-      rect.height > 0 &&
-      style.display !== 'none' &&
-      style.visibility !== 'hidden' &&
-      Number(style.opacity || '1') > 0
-    );
-  }
-
-  function roleOf(element) {
-    const explicit = element.getAttribute('role');
-    if (explicit) {
-      return explicit.toLowerCase();
-    }
-    const tag = element.tagName.toLowerCase();
-    if (tag === 'a' && element.getAttribute('href')) {
-      return 'link';
-    }
-    if (tag === 'button') {
-      return 'button';
-    }
-    if (['input', 'select', 'textarea'].includes(tag)) {
-      return 'form-control';
-    }
-    if (tag === 'summary') {
-      return 'button';
-    }
-    if (['header', 'nav', 'main', 'footer', 'section', 'aside'].includes(tag)) {
-      return tag;
-    }
-    return 'group';
-  }
-
-  function isInteractive(element) {
-    const role = roleOf(element);
-    const tag = element.tagName.toLowerCase();
-    return role === 'link' || role === 'button' || role === 'form-control' || ['a', 'button', 'input', 'select', 'textarea', 'summary'].includes(tag);
-  }
-
-  function classTokens(element) {
-    return [...element.classList].filter((token) => meaningfulHookPattern.test(token)).sort();
-  }
-
-  function childClassTokens(element) {
-    return [...element.querySelectorAll('[class]')]
-      .flatMap((child) => classTokens(child))
-      .filter((token, index, values) => values.indexOf(token) === index)
-      .sort();
-  }
-
-  function regionOf(element) {
-    const region = element.closest('footer,header,nav,main,section,aside,[role=banner],[role=navigation],[role=main],[role=contentinfo]');
-    if (!region) {
-      return 'body';
-    }
-    const role = roleOf(region);
-    if (role === 'banner') {
-      return 'header';
-    }
-    if (role === 'navigation') {
-      return 'nav';
-    }
-    if (role === 'contentinfo') {
-      return 'footer';
-    }
-    return region.tagName.toLowerCase();
-  }
-
-  function boxOf(element) {
-    const rect = element.getBoundingClientRect();
-    return {
-      x: Math.round(rect.x),
-      y: Math.round(rect.y),
-      width: Math.round(rect.width),
-      height: Math.round(rect.height),
-      area: Math.round(rect.width * rect.height),
-    };
-  }
-
-  function containsLogo(element) {
-    return Boolean(
-      element.querySelector('img,svg,picture') ||
-        /\b(?:brand|logo|wordmark)\b/i.test(element.className || '') ||
-        /\b(?:logo|brand|wordmark)\b/i.test(element.getAttribute('aria-label') || '')
-    );
-  }
-
-  function containsWordmark(element) {
-    return Boolean(
-      /\bwordmark\b/i.test(element.className || '') ||
-        element.querySelector('[class*=wordmark]') ||
-        normalizeText(element.textContent).length > 0
-    );
-  }
-
-  function clickableDescendants(element) {
-    return [...element.querySelectorAll('a[href],button,[role=button],[role=link],input,select,textarea,summary')].filter(visible);
-  }
-
-  function visualPartCount(element) {
-    return [...element.children].filter((child) => visible(child)).length;
-  }
-
-  function elementSummary(element, extra = {}) {
-    const clickable = clickableDescendants(element);
-    const box = boxOf(element);
-    return {
-      tag: element.tagName.toLowerCase(),
-      role: roleOf(element),
-      text: normalizeText(element.textContent),
-      href: element.getAttribute('href') || '',
-      own_classes: classTokens(element),
-      child_classes: childClassTokens(element),
-      contains_logo: containsLogo(element),
-      contains_wordmark: containsWordmark(element),
-      contains_image: Boolean(element.querySelector('img,picture,video')),
-      contains_svg: Boolean(element.querySelector('svg')),
-      clickable_descendant_count: clickable.length + (isInteractive(element) ? 1 : 0),
-      child_visual_part_count: visualPartCount(element),
-      wraps_multiple_visual_parts: visualPartCount(element) >= 2,
-      ancestor_region: regionOf(element),
-      bounding_box: box,
-      ...extra,
-    };
-  }
-
-  function conceptForElement(element) {
-    const haystack = [element.className || '', normalizeText(element.textContent), element.getAttribute('aria-label') || ''].join(' ');
-    const match = haystack.match(meaningfulHookPattern);
-    return match ? match[1].toLowerCase() : '';
-  }
-
-  const landmarks = {};
-  for (const [name, selector] of Object.entries(landmarkSelectors)) {
-    const matches = [...document.querySelectorAll(selector)].filter(visible);
-    landmarks[name] = {
-      count: matches.length,
-      visible_count: matches.length,
-      first_text: normalizeText(matches[0]?.textContent || ''),
-    };
-  }
-
-  const classOwners = [];
-  for (const element of [...document.querySelectorAll('[class]')].filter(visible)) {
-    const classes = classTokens(element);
-    if (!classes.length) {
-      continue;
-    }
-    classOwners.push(elementSummary(element, { selector_signature: `.${classes[0]}`, concept: conceptForElement(element) }));
-  }
-
-  const interactions = [...document.querySelectorAll('a[href],button,[role=button],[role=link],input,select,textarea,summary')]
-    .filter(visible)
-    .map((element) => elementSummary(element, { concept: conceptForElement(element) }));
-
-  const regions = {};
-  for (const name of ['header', 'nav', 'main', 'footer', 'section', 'aside', 'body']) {
-    regions[name] = {
-      link_count: 0,
-      button_count: 0,
-      clickable_area: 0,
-      media_count: 0,
-      brand_present: false,
-      logo_present: false,
-      text: '',
-    };
-  }
-
-  for (const interaction of interactions) {
-    const region = regions[interaction.ancestor_region] || regions.body;
-    if (interaction.role === 'link') {
-      region.link_count++;
-    }
-    if (interaction.role === 'button') {
-      region.button_count++;
-    }
-    region.clickable_area += interaction.bounding_box.area;
-  }
-
-  for (const [name, region] of Object.entries(regions)) {
-    const root = name === 'body' ? document.body : document.querySelector(name);
-    if (!root) {
-      continue;
-    }
-    region.media_count = root.querySelectorAll('img,svg,picture,video').length;
-    region.brand_present = Boolean(root.querySelector('[class*=brand],[class*=wordmark]'));
-    region.logo_present = Boolean(root.querySelector('img,svg,picture,[class*=logo]'));
-    region.text = normalizeText(root.textContent, 260);
-  }
-
-  const repeated = {
-    card: document.querySelectorAll('[class*=card],article').length,
-    list_item: document.querySelectorAll('li').length,
-    feature: document.querySelectorAll('[class*=feature]').length,
-    plan: document.querySelectorAll('[class*=plan],[class*=price]').length,
-    testimonial: document.querySelectorAll('[class*=testimonial]').length,
-  };
-
-  const selector_groups = [];
-  for (const group of groups || []) {
-    const selectors = [];
-    for (const selector of group.selectors || []) {
-      try {
-        const matches = [...document.querySelectorAll(selector)].filter(visible);
-        selectors.push({
-          selector,
-          count: matches.length,
-          first: matches[0] ? elementSummary(matches[0], { concept: conceptForElement(matches[0]) }) : null,
-        });
-      } catch (error) {
-        selectors.push({ selector, count: 0, first: null, error: error instanceof Error ? error.message : String(error) });
-      }
-    }
-    selector_groups.push({ name: group.name, selectors });
-  }
-
-  return {
-    url: window.location.href,
-    title: document.title,
-    landmarks,
-    class_owners: classOwners,
-    interactions,
-    regions,
-    repeated,
-    selector_groups,
-  };
-}
-
-async function evaluateSemanticSurface(page, groups) {
-  return page.evaluate(semanticFingerprintExtractor, groups);
-}
-
-function semanticPrimaryClassKey(owner) {
-  return owner?.own_classes?.[0] || '';
-}
-
-function semanticRole(owner) {
-  return owner?.role || owner?.tag || '';
-}
-
-function semanticTextTokens(value) {
-  return String(value || '')
-    .toLowerCase()
-    .replace(/[^a-z0-9\s]+/g, ' ')
-    .split(/\s+/)
-    .filter((token) => token.length >= 3);
-}
-
-function semanticHasMaterialRepeatedDelta(sourceCount, frontendCount) {
-  if (sourceCount < 3 && frontendCount < 3) {
-    return false;
-  }
-  return Math.abs(sourceCount - frontendCount) >= Math.max(3, Math.ceil(sourceCount * 0.35));
-}
-
-function semanticAllowsNavigationClassRoleChange(sourceOwner, frontendOwner) {
-  const sourceRole = semanticRole(sourceOwner);
-  const frontendRole = semanticRole(frontendOwner);
-  const key = semanticPrimaryClassKey(sourceOwner);
-  if (!/nav|menu/i.test(key) || !['group', 'list'].includes(sourceRole) || frontendRole !== 'nav') {
-    return false;
-  }
-
-  return (
-    Number(sourceOwner.clickable_descendant_count || 0) === Number(frontendOwner.clickable_descendant_count || 0) &&
-    semanticTextTokens(sourceOwner.text).every((token) => new Set(semanticTextTokens(frontendOwner.text)).has(token))
-  );
-}
-
-function semanticMismatch(type, reason, source, frontend, extra = {}) {
-  return {
-    type,
-    reason,
-    region: source?.ancestor_region || frontend?.ancestor_region || extra.region || '',
-    concept: source?.concept || frontend?.concept || extra.concept || '',
-    selector_signature: source?.selector_signature || frontend?.selector_signature || extra.selector_signature || '',
-    source,
-    generated: frontend,
-    ...extra,
-  };
-}
-
-function compareSemanticFingerprints(source, frontend) {
-  const mismatches = [];
-  const optionalSelectorAbsences = [];
-  const counts = {
-    role_mismatch_count: 0,
-    class_owner_changed_count: 0,
-    interaction_group_split_count: 0,
-    interaction_group_merged_count: 0,
-    link_text_delta_count: 0,
-    landmark_mismatch_count: 0,
-    repeated_count_delta_count: 0,
-    brand_logo_missing_count: 0,
-  };
-
-  for (const landmark of ['header', 'nav', 'main', 'footer']) {
-    const sourceCount = Number(source?.landmarks?.[landmark]?.visible_count || 0);
-    const frontendCount = Number(frontend?.landmarks?.[landmark]?.visible_count || 0);
-    if (sourceCount > 0 && frontendCount === 0) {
-      counts.landmark_mismatch_count++;
-      mismatches.push(
-        semanticMismatch('landmark', 'landmark_disappeared', source?.landmarks?.[landmark], frontend?.landmarks?.[landmark], {
-          region: landmark,
-          concept: landmark,
-        })
-      );
-    }
-  }
-
-  const frontendOwnersByClass = new Map();
-  for (const owner of frontend?.class_owners || []) {
-    const key = semanticPrimaryClassKey(owner);
-    if (key && !frontendOwnersByClass.has(key)) {
-      frontendOwnersByClass.set(key, owner);
-    }
-  }
-
-  for (const sourceOwner of source?.class_owners || []) {
-    const key = semanticPrimaryClassKey(sourceOwner);
-    const frontendOwner = key ? frontendOwnersByClass.get(key) : null;
-    if (!frontendOwner) {
-      continue;
-    }
-
-    const sourceRole = semanticRole(sourceOwner);
-    const frontendRole = semanticRole(frontendOwner);
-    const sourceInteractive = ['link', 'button', 'form-control'].includes(sourceRole);
-    const frontendInteractive = ['link', 'button', 'form-control'].includes(frontendRole);
-    const roleChanged = sourceRole !== frontendRole;
-    const sourceClickable = Number(sourceOwner.clickable_descendant_count || 0);
-    const frontendClickable = Number(frontendOwner.clickable_descendant_count || 0);
-
-    if (sourceRole === 'link' && frontendRole !== 'link') {
-      counts.role_mismatch_count++;
-      counts.class_owner_changed_count++;
-      if (frontendClickable > sourceClickable) {
-        counts.interaction_group_split_count++;
-      }
-      mismatches.push(
-        semanticMismatch('class_owner', 'classed_link_became_non_link', sourceOwner, frontendOwner, {
-          selector_signature: `.${key}`,
-        })
-      );
-      continue;
-    }
-
-    if (
-      roleChanged &&
-      !semanticAllowsNavigationClassRoleChange(sourceOwner, frontendOwner) &&
-      (sourceInteractive || frontendInteractive || sourceOwner.concept || frontendOwner.concept)
-    ) {
-      counts.role_mismatch_count++;
-      counts.class_owner_changed_count++;
-      mismatches.push(
-        semanticMismatch('class_owner', 'meaningful_class_moved_role', sourceOwner, frontendOwner, {
-          selector_signature: `.${key}`,
-        })
-      );
-    }
-
-    if (sourceInteractive && frontendClickable > sourceClickable + 1) {
-      counts.interaction_group_split_count++;
-      mismatches.push(
-        semanticMismatch('interaction_group', 'source_interaction_group_split', sourceOwner, frontendOwner, {
-          selector_signature: `.${key}`,
-        })
-      );
-    } else if (!sourceInteractive && frontendInteractive && sourceClickable > frontendClickable + 1) {
-      counts.interaction_group_merged_count++;
-      mismatches.push(
-        semanticMismatch('interaction_group', 'source_interaction_group_merged', sourceOwner, frontendOwner, {
-          selector_signature: `.${key}`,
-        })
-      );
-    }
-
-    const sourceTokens = semanticTextTokens(sourceOwner.text);
-    const frontendTokens = new Set(semanticTextTokens(frontendOwner.text));
-    const missingTokens = sourceTokens.filter((token) => !frontendTokens.has(token));
-    if ((sourceRole === 'link' || sourceRole === 'button') && sourceTokens.length && missingTokens.length === sourceTokens.length) {
-      counts.link_text_delta_count++;
-      mismatches.push(
-        semanticMismatch('interaction_text', 'link_or_button_text_disappeared', sourceOwner, frontendOwner, {
-          selector_signature: `.${key}`,
-          missing_text_tokens: missingTokens,
-        })
-      );
-    }
-  }
-
-  for (const [regionName, sourceRegion] of Object.entries(source?.regions || {})) {
-    const frontendRegion = frontend?.regions?.[regionName];
-    if (!frontendRegion) {
-      continue;
-    }
-    const sourceHasBrandLogo = sourceRegion.brand_present && sourceRegion.logo_present;
-    const frontendHasBrandLogo = frontendRegion.brand_present && frontendRegion.logo_present;
-    if (['header', 'footer'].includes(regionName) && sourceHasBrandLogo && !frontendHasBrandLogo) {
-      counts.brand_logo_missing_count++;
-      mismatches.push(
-        semanticMismatch('brand_media', 'brand_or_logo_image_disappeared', sourceRegion, frontendRegion, {
-          region: regionName,
-          concept: 'brand',
-        })
-      );
-    }
-  }
-
-  for (const [name, sourceCount] of Object.entries(source?.repeated || {})) {
-    const frontendCount = Number(frontend?.repeated?.[name] || 0);
-    if (semanticHasMaterialRepeatedDelta(Number(sourceCount || 0), frontendCount)) {
-      counts.repeated_count_delta_count++;
-      mismatches.push(
-        semanticMismatch('repeated_structure', 'repeated_structure_count_changed_materially', { count: sourceCount }, { count: frontendCount }, {
-          concept: name,
-        })
-      );
-    }
-  }
-
-  const frontendSelectorGroups = new Map((frontend?.selector_groups || []).map((group) => [group.name, group]));
-  for (const sourceGroup of source?.selector_groups || []) {
-    const frontendGroup = frontendSelectorGroups.get(sourceGroup.name);
-    if (!frontendGroup) {
-      continue;
-    }
-    const frontendSelectors = new Map((frontendGroup.selectors || []).map((selector) => [selector.selector, selector]));
-    for (const sourceSelector of sourceGroup.selectors || []) {
-      const frontendSelector = frontendSelectors.get(sourceSelector.selector);
-      if (!frontendSelector) {
-        continue;
-      }
-      if (sourceSelector.count === 0 && frontendSelector.count === 0) {
-        optionalSelectorAbsences.push({ group: sourceGroup.name, selector: sourceSelector.selector });
-      }
-    }
-  }
-
-  const sourceTotals = semanticSurfaceTotals(source);
-  const frontendTotals = semanticSurfaceTotals(frontend);
-  const sourceArea = sourceTotals.clickable_area;
-  const areaDeltaRatio = sourceArea > 0 ? Math.abs(sourceArea - frontendTotals.clickable_area) / sourceArea : 0;
-
-  return {
-    mismatch_count: mismatches.length,
-    ...counts,
-    region_link_count_delta: frontendTotals.region_link_count - sourceTotals.region_link_count,
-    clickable_area_delta_ratio: Number(areaDeltaRatio.toFixed(4)),
-    mismatches,
-    optional_selector_absences: optionalSelectorAbsences,
-  };
-}
-
-function buildSemanticArtifact(results, artifactPath) {
-  const mismatches = [];
-  const targets = [];
-  for (const result of results) {
-    const comparison = result.comparison || {};
-    for (const mismatch of comparison.mismatches || []) {
-      mismatches.push({
-        target: result.source_filename || String(result.wordpress_page_id || ''),
-        ...mismatch,
-      });
-    }
-    targets.push({
-      source_filename: result.source_filename || '',
-      wordpress_page_id: result.wordpress_page_id || null,
-      mismatch_count: Number(comparison.mismatch_count || 0),
-      role_mismatch_count: Number(comparison.role_mismatch_count || 0),
-      class_owner_changed_count: Number(comparison.class_owner_changed_count || 0),
-      interaction_group_split_count: Number(comparison.interaction_group_split_count || 0),
-      interaction_group_merged_count: Number(comparison.interaction_group_merged_count || 0),
-      link_text_delta_count: Number(comparison.link_text_delta_count || 0),
-      landmark_mismatch_count: Number(comparison.landmark_mismatch_count || 0),
-      repeated_count_delta_count: Number(comparison.repeated_count_delta_count || 0),
-      brand_logo_missing_count: Number(comparison.brand_logo_missing_count || 0),
-      region_link_count_delta: Number(comparison.region_link_count_delta || 0),
-      clickable_area_delta_ratio: Number(comparison.clickable_area_delta_ratio || 0),
-      optional_selector_absent_count: Number(comparison.optional_selector_absences?.length || 0),
-    });
-  }
-
-  return {
-    artifact: artifactPath,
-    target_count: results.length,
-    mismatch_count: mismatches.length,
-    targets,
-    mismatches,
-    results,
-  };
-}
-
 export async function compareSemanticFidelity(importReport, artifactDir, sitePath) {
-  const targets = semanticComparisonTargets(importReport);
-  if (!targets.length) {
-    const semanticDir = path.join(artifactDir, 'semantic-comparisons');
-    const artifactPath = path.join(semanticDir, 'semantic-fidelity-skipped.json');
-    const reason = importReport?.error || 'No semantic or visual fidelity comparison targets found.';
-    const artifact = {
-      ...buildSemanticArtifact([], artifactPath),
-      skipped: true,
-      reason,
-    };
-    await mkdir(semanticDir, { recursive: true });
-    await writeFile(artifactPath, JSON.stringify(artifact, null, 2));
-    return emptySemanticComparison(reason, artifactPath, artifact);
-  }
-
-  const playwrightPackage = path.join(STUDIO_PATH, 'node_modules/@playwright/test');
-  const { chromium } = requireFromBench(playwrightPackage);
-  const semanticDir = path.join(artifactDir, 'semantic-comparisons');
-  await mkdir(semanticDir, { recursive: true });
-
-  let browser;
-  const results = [];
-
-  try {
-    browser = await chromium.launch({ args: ['--ignore-certificate-errors'] });
-    for (const [index, target] of targets.entries()) {
-      const targetSlug = safeSlug(target.source_filename || target.wordpress_page_id, `target-${index + 1}`);
-      const selectorGroups = semanticTargetSelectorGroups(target);
-      const result = {
-        source_filename: target.source_filename || '',
-        wordpress_page_id: target.wordpress_page_id || null,
-        generated_template: target.generated_template || '',
-        generated_pattern: target.generated_pattern || '',
-        comparison_hooks: target.comparison_hooks || {},
-        surfaces: {},
-        comparison: null,
-        errors: [],
-      };
-
-      for (const surface of ['source_static', 'wordpress_frontend']) {
-        const url = surfaceUrl(target, surface, importReport.reportPath, sitePath);
-        const page = await browser.newPage({ ignoreHTTPSErrors: true, viewport: VISUAL_VIEWPORT });
-        try {
-          if (!url) {
-            throw new Error(`Missing ${surface} render URL.`);
-          }
-          await loadVisualSurface(page, url);
-          result.surfaces[surface] = {
-            url,
-            fingerprint: await evaluateSemanticSurface(page, selectorGroups),
-          };
-        } catch (error) {
-          const message = error instanceof Error ? error.message : String(error);
-          result.errors.push(`${surface}: ${message}`);
-          result.surfaces[surface] = { url, fingerprint: null, error: message };
-        } finally {
-          await page.close();
-        }
-      }
-
-      if (result.surfaces.source_static?.fingerprint && result.surfaces.wordpress_frontend?.fingerprint) {
-        result.comparison = compareSemanticFingerprints(
-          result.surfaces.source_static.fingerprint,
-          result.surfaces.wordpress_frontend.fingerprint
-        );
-      } else {
-        result.comparison = compareSemanticFingerprints({}, {});
-      }
-
-      await writeFile(path.join(semanticDir, `${targetSlug}-semantic-fingerprint.json`), JSON.stringify(result, null, 2));
-      results.push(result);
-    }
-  } catch (error) {
-    return emptySemanticComparison(`Semantic comparison failed: ${error instanceof Error ? error.message : String(error)}`);
-  } finally {
-    if (browser) {
-      await browser.close();
-    }
-  }
-
-  const totals = results.reduce(
-    (summary, result) => {
-      const comparison = result.comparison || {};
-      summary.error_count += result.errors.length;
-      summary.checked_target_count += result.errors.length === 0 ? 1 : 0;
-      summary.mismatch_count += Number(comparison.mismatch_count || 0);
-      summary.role_mismatch_count += Number(comparison.role_mismatch_count || 0);
-      summary.class_owner_changed_count += Number(comparison.class_owner_changed_count || 0);
-      summary.interaction_group_split_count += Number(comparison.interaction_group_split_count || 0);
-      summary.interaction_group_merged_count += Number(comparison.interaction_group_merged_count || 0);
-      summary.link_text_delta_count += Number(comparison.link_text_delta_count || 0);
-      summary.region_link_count_delta += Math.abs(Number(comparison.region_link_count_delta || 0));
-      summary.clickable_area_delta_ratio += Number(comparison.clickable_area_delta_ratio || 0);
-      summary.landmark_mismatch_count += Number(comparison.landmark_mismatch_count || 0);
-      summary.repeated_count_delta_count += Number(comparison.repeated_count_delta_count || 0);
-      summary.brand_logo_missing_count += Number(comparison.brand_logo_missing_count || 0);
-      return summary;
-    },
-    {
-      target_count: targets.length,
-      checked_target_count: 0,
-      error_count: 0,
-      mismatch_count: 0,
-      role_mismatch_count: 0,
-      class_owner_changed_count: 0,
-      interaction_group_split_count: 0,
-      interaction_group_merged_count: 0,
-      link_text_delta_count: 0,
-      region_link_count_delta: 0,
-      clickable_area_delta_ratio: 0,
-      landmark_mismatch_count: 0,
-      repeated_count_delta_count: 0,
-      brand_logo_missing_count: 0,
-    }
-  );
-
-  if (results.length) {
-    totals.clickable_area_delta_ratio = Number((totals.clickable_area_delta_ratio / results.length).toFixed(4));
-  }
-
-  const artifactPath = path.join(semanticDir, 'semantic-fidelity.json');
-  const artifact = buildSemanticArtifact(results, artifactPath);
-  await writeFile(artifactPath, JSON.stringify(artifact, null, 2));
-
-  return {
-    ...totals,
-    artifact_dir: semanticDir,
-    artifact: artifactPath,
-    diagnostics: artifact,
-    results,
-  };
-}
-
-async function readIfExists(file) {
-  try {
-    return await readFile(file, 'utf8');
-  } catch {
-    return '';
-  }
-}
-
-async function collectCssFiles(root, maxFiles = 40) {
-  const files = [];
-
-  async function visit(dir) {
-    if (files.length >= maxFiles) {
-      return;
-    }
-    let entries = [];
-    try {
-      entries = await readdir(dir, { withFileTypes: true });
-    } catch {
-      return;
-    }
-
-    for (const entry of entries) {
-      if (files.length >= maxFiles) {
-        return;
-      }
-      if (entry.name === 'node_modules' || entry.name === '.git') {
-        continue;
-      }
-      const absolute = path.join(dir, entry.name);
-      if (entry.isDirectory()) {
-        await visit(absolute);
-      } else if (entry.isFile() && entry.name.endsWith('.css')) {
-        files.push(absolute);
-      }
-    }
-  }
-
-  await visit(root);
-  return files;
-}
-
-export function reportedFreeformBlockCount(importReport) {
-  const quality = importReport?.report?.quality || {};
-  const generatedTheme = importReport?.report?.generated_theme || {};
-  const candidates = [
-    quality.freeform_block_count,
-    quality.freeform_blocks,
-    generatedTheme.freeform_block_count,
-    generatedTheme.freeform_blocks,
-  ];
-
-  for (const candidate of candidates) {
-    const value = Number(candidate);
-    if (Number.isFinite(value)) {
-      return value;
-    }
-  }
-
-  return 0;
-}
-
-function freeformDiagnostics(documents) {
-  const results = [];
-  let count = 0;
-
-  for (const document of documents) {
-    const documentCount = countRegex(document.content, /<!--\s+wp:freeform\b/gi);
-    if (documentCount === 0) {
-      continue;
-    }
-    count += documentCount;
-    results.push({
-      source: document.source,
-      freeform_block_count: documentCount,
-    });
-  }
-
-  return { count, results };
-}
-
-function splitSelectorList(selector) {
-  const selectors = [];
-  let current = '';
-  let depth = 0;
-  let quote = '';
-
-  for (const char of String(selector || '')) {
-    if (quote) {
-      current += char;
-      if (char === quote) {
-        quote = '';
-      }
-      continue;
-    }
-
-    if (char === '"' || char === "'") {
-      quote = char;
-      current += char;
-      continue;
-    }
-
-    if (char === '(' || char === '[') {
-      depth += 1;
-      current += char;
-      continue;
-    }
-    if (char === ')' || char === ']') {
-      depth = Math.max(0, depth - 1);
-      current += char;
-      continue;
-    }
-
-    if (char === ',' && depth === 0) {
-      const value = current.trim();
-      if (value) {
-        selectors.push(value);
-      }
-      current = '';
-      continue;
-    }
-
-    current += char;
-  }
-
-  const value = current.trim();
-  if (value) {
-    selectors.push(value);
-  }
-
-  return selectors;
-}
-
-function selectorCoverageTokens(selector) {
-  const tokens = new Set();
-  for (const match of String(selector || '').matchAll(/\.(-?[_a-zA-Z]+[_a-zA-Z0-9-]*)/g)) {
-    const name = match[1];
-    if (!/^(?:editor-styles-wrapper|block-editor|wp-admin|is-root-container)$/.test(name)) {
-      tokens.add(`.${name}`);
-    }
-  }
-  for (const match of String(selector || '').matchAll(/\[\s*([^\]\s~|^$*=]+)/g)) {
-    tokens.add(`[${match[1].toLowerCase()}]`);
-  }
-  return tokens;
-}
-
-function editorOverrideCoversHiddenSelector(hiddenSelector, overrideSelector) {
-  const hiddenTokens = selectorCoverageTokens(hiddenSelector);
-  if (hiddenTokens.size === 0) {
-    return false;
-  }
-
-  const overrideTokens = selectorCoverageTokens(overrideSelector);
-  if (overrideTokens.size === 0) {
-    return false;
-  }
-
-  for (const token of overrideTokens) {
-    if (!hiddenTokens.has(token)) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-export function hiddenEditorContentDiagnostics(cssFiles) {
-  const hiddenRules = [];
-  const editorOverrideRules = [];
-  const hiddenSelectors = [];
-  const editorOverrideSelectors = [];
-  const revealSelectorPattern = /(?:^|[\s.#:[>,])(?:js[-_])?(?:reveal|revealed|aos|fade[-_]in|animate[-_]on[-_]scroll|scroll[-_]reveal)\b|\[data[-_](?:reveal|aos|animate)/i;
-  const hiddenSelectorPattern = /(?:^|[\s.#:[>,])hidden\b/i;
-  const hiddenDeclarationPattern = /(?:opacity\s*:\s*0(?:\.0+)?\b|visibility\s*:\s*hidden\b|display\s*:\s*none\b)/i;
-  const visibleDeclarationPattern = /(?:opacity\s*:\s*1(?:\.0+)?\b|visibility\s*:\s*visible\b|display\s*:\s*(?:block|grid|flex|contents)\b)/i;
-  const editorSelectorPattern = /(?:editor-styles-wrapper|block-editor|wp-admin|is-root-container)/i;
-
-  for (const file of cssFiles) {
-    const content = typeof file.content === 'string' ? file.content : '';
-    for (const match of content.matchAll(/([^{}]+)\{([^{}]*)\}/g)) {
-      const selector = match[1].trim().replace(/\s+/g, ' ');
-      const declarations = match[2].trim().replace(/\s+/g, ' ');
-      const isRevealSelector = revealSelectorPattern.test(selector);
-      const isBroadEditorHiddenOverride =
-        editorSelectorPattern.test(selector) && hiddenSelectorPattern.test(selector) && visibleDeclarationPattern.test(declarations);
-      if (!isRevealSelector && !isBroadEditorHiddenOverride) {
-        continue;
-      }
-
-      const rule = {
-        source: file.source,
-        selector,
-        declarations: declarations.slice(0, 300),
-      };
-
-      const selectorList = splitSelectorList(selector);
-
-      if (isRevealSelector && hiddenDeclarationPattern.test(declarations)) {
-        hiddenRules.push(rule);
-        for (const selectorItem of selectorList) {
-          if (revealSelectorPattern.test(selectorItem)) {
-            hiddenSelectors.push({ ...rule, selector: selectorItem });
-          }
-        }
-      }
-      if (editorSelectorPattern.test(selector) && visibleDeclarationPattern.test(declarations)) {
-        editorOverrideRules.push(rule);
-        for (const selectorItem of selectorList) {
-          if (editorSelectorPattern.test(selectorItem)) {
-            editorOverrideSelectors.push({ ...rule, selector: selectorItem });
-          }
-        }
-      }
-    }
-  }
-
-  const missingEditorOverrideSelectors = hiddenSelectors.filter(
-    (hiddenRule) =>
-      !editorOverrideSelectors.some((editorRule) =>
-        editorOverrideCoversHiddenSelector(hiddenRule.selector, editorRule.selector)
-      )
-  );
-
-  return {
-    hidden_rule_count: hiddenRules.length,
-    editor_override_rule_count: editorOverrideRules.length,
-    missing_editor_override_count: missingEditorOverrideSelectors.length,
-    hidden_rules: hiddenRules.slice(0, 25),
-    editor_override_rules: editorOverrideRules.slice(0, 25),
-    missing_editor_override_rules: missingEditorOverrideSelectors.slice(0, 25),
-  };
-}
-
-function cssRules(cssFiles) {
-  const rules = [];
-  for (const file of cssFiles) {
-    const content = typeof file.content === 'string' ? file.content : '';
-    for (const match of content.matchAll(/([^{}]+)\{([^{}]*)\}/g)) {
-      const selector = match[1].trim().replace(/\s+/g, ' ');
-      if (!selector || selector.startsWith('@')) {
-        continue;
-      }
-      rules.push({
-        source: file.source,
-        selector,
-        declarations: match[2].trim().replace(/\s+/g, ' '),
-      });
-    }
-  }
-  return rules;
-}
-
-function structuralSelectorTarget(selector) {
-  const normalizedSelector = String(selector || '')
-    .replace(/::?[a-z-]+(?:\([^)]*\))?/gi, '')
-    .trim();
-  const match = normalizedSelector.match(
-    /(?:^|[\s>+~])(?<tag>header|main|nav|footer|section|article|aside)(?<compound>(?:[#.][-_a-zA-Z0-9]+)*)/i
-  );
-  if (!match?.groups) {
-    return null;
-  }
-
-  const compound = match.groups.compound || '';
-  const id = compound.match(/#([-_a-zA-Z0-9]+)/)?.[1] || '';
-  const classes = [...compound.matchAll(/\.([-_a-zA-Z0-9]+)/g)].map((classMatch) => classMatch[1]);
-  const text = `${match.groups.tag} ${id} ${classes.join(' ')} ${selector}`;
-  const heroLike = /\bhero\b/i.test(text);
-  if (!id && classes.length === 0) {
-    return null;
-  }
-
-  return {
-    tag: match.groups.tag.toLowerCase(),
-    id,
-    classes,
-    hero_like: heroLike,
-  };
-}
-
-function hasMaterialStructuralLayout(declarations, target) {
-  const layoutPattern = /(?:display\s*:\s*(?:grid|flex)|grid-template|flex(?:-direction|-wrap)?\s*:|align-items\s*:|justify-content\s*:|min-height\s*:|height\s*:\s*(?:\d+(?:vh|dvh|svh)|100%)|padding(?:-[a-z]+)?\s*:|background(?:-|\s*:)|position\s*:\s*(?:relative|absolute|fixed|sticky))/i;
-  return target?.hero_like === true && layoutPattern.test(String(declarations || ''));
-}
-
-function htmlAttributeTokens(attributes, name) {
-  const values = [];
-  const pattern = new RegExp(`\\b${name}\\s*=\\s*(["'])(.*?)\\1`, 'gi');
-  for (const match of String(attributes || '').matchAll(pattern)) {
-    values.push(...String(match[2] || '').split(/\s+/).filter(Boolean));
-  }
-  return values;
-}
-
-function markupHasId(markup, id) {
-  if (!id) {
-    return true;
-  }
-  const escaped = id.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-  return (
-    new RegExp(`\\bid\\s*=\\s*(["'])${escaped}\\1`, 'i').test(markup) ||
-    new RegExp(`"anchor"\\s*:\\s*"${escaped}"`, 'i').test(markup)
-  );
-}
-
-function markupHasClasses(markup, classes) {
-  const classValues = [];
-  const pattern = /\bclass(?:Name)?\s*=\s*(["'])(.*?)\1/gi;
-  for (const match of String(markup || '').matchAll(pattern)) {
-    classValues.push(...String(match[2] || '').split(/\s+/).filter(Boolean));
-  }
-  return classes.every((className) => classValues.includes(className));
-}
-
-function markupHasStructuralTarget(markup, target) {
-  const tagPattern = new RegExp(`<${target.tag}\\b([^>]*)>`, 'gi');
-  for (const match of String(markup || '').matchAll(tagPattern)) {
-    const attributes = match[1] || '';
-    const ids = htmlAttributeTokens(attributes, 'id');
-    const classes = htmlAttributeTokens(attributes, 'class');
-    const idMatches = !target.id || ids.includes(target.id);
-    const classesMatch = target.classes.every((className) => classes.includes(className));
-    if (idMatches && classesMatch) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function structuralSelectorDriftReason(markup, target) {
-  if (target.id && !markupHasId(markup, target.id)) {
-    return 'missing_generated_dom_id';
-  }
-  if (target.classes.length > 0 && !markupHasClasses(markup, target.classes)) {
-    return 'missing_generated_dom_class';
-  }
-  return 'generated_dom_tag_drift';
-}
-
-export function structuralSelectorDriftDiagnostics(cssFiles, generatedDocuments) {
-  const generatedMarkup = generatedDocuments.map((document) => document.content || '').join('\n');
-  const materialSelectors = [];
-  const missingSelectors = [];
-
-  for (const rule of cssRules(cssFiles)) {
-    for (const selector of splitSelectorList(rule.selector)) {
-      const target = structuralSelectorTarget(selector);
-      if (!target || !hasMaterialStructuralLayout(rule.declarations, target)) {
-        continue;
-      }
-
-      const diagnostic = {
-        source: rule.source,
-        selector,
-        expected_tag: target.tag,
-        expected_id: target.id,
-        expected_classes: target.classes,
-        declarations: rule.declarations.slice(0, 300),
-      };
-      materialSelectors.push(diagnostic);
-      if (!markupHasStructuralTarget(generatedMarkup, target)) {
-        missingSelectors.push({
-          ...diagnostic,
-          reason: structuralSelectorDriftReason(generatedMarkup, target),
-        });
-      }
-    }
-  }
-
-  return {
-    source_structural_selector_count: materialSelectors.length,
-    missing_structural_selector_count: missingSelectors.length,
-    source_structural_selectors: materialSelectors.slice(0, 25),
-    missing_structural_selectors: missingSelectors.slice(0, 25),
-  };
-}
-
-export async function collectGeneratedThemeUxGates(sitePath, importReport, artifactDir) {
-  const { themeRoot, themeSlug, error } = await collectLatestGeneratedTheme(sitePath);
-  const artifactPath = path.join(artifactDir, 'generated-theme-ux-gates.json');
-  const reasons = [];
-
-  if (!themeRoot) {
-    const skipped = {
-      skipped: true,
-      error,
-      theme_slug: themeSlug,
-      artifact: artifactPath,
-      generated_theme_ux_quality_pass: false,
-      generated_theme_ux_quality_failure_count: 1,
-      generated_theme_ux_quality_failure_reasons: ['missing_generated_theme'],
-    };
-    await writeFile(artifactPath, JSON.stringify(skipped, null, 2));
-    return skipped;
-  }
-
-  const documents = await collectThemeBlockDocuments(sitePath);
-  const freeform = freeformDiagnostics(documents);
-  const importerFreeformCount = reportedFreeformBlockCount(importReport);
-  const cssFiles = await collectCssFiles(themeRoot);
-  const cssFileContents = [];
-  for (const file of cssFiles) {
-    cssFileContents.push({ source: path.relative(themeRoot, file), content: await readIfExists(file) });
-  }
-  const hiddenEditorContent = hiddenEditorContentDiagnostics(cssFileContents);
-  const sourceRoot = path.join(sitePath, 'tmp/static-site');
-  const sourceCssFiles = await collectCssFiles(sourceRoot);
-  const sourceCssFileContents = [];
-  for (const file of sourceCssFiles) {
-    sourceCssFileContents.push({ source: path.relative(sourceRoot, file), content: await readIfExists(file) });
-  }
-  const structuralSelectorDrift = structuralSelectorDriftDiagnostics(sourceCssFileContents, documents);
-
-  if (freeform.count !== importerFreeformCount) {
-    reasons.push('freeform_report_count_mismatch');
-  }
-  if (hiddenEditorContent.missing_editor_override_count > 0) {
-    reasons.push('css_hidden_editor_content_without_override');
-  }
-  if (structuralSelectorDrift.missing_structural_selector_count > 0) {
-    reasons.push('source_css_structural_selector_drift');
-  }
-
-  const diagnostics = {
-    skipped: false,
-    theme_slug: themeSlug,
-    theme_root: themeRoot,
-    document_count: documents.length,
-    css_file_count: cssFiles.length,
-    actual_freeform_block_count: freeform.count,
-    importer_freeform_block_count: importerFreeformCount,
-    freeform_report_mismatch_count: freeform.count === importerFreeformCount ? 0 : 1,
-    freeform_documents: freeform.results,
-    css_hidden_editor_content_count: hiddenEditorContent.hidden_rule_count,
-    css_editor_reveal_override_count: hiddenEditorContent.editor_override_rule_count,
-    css_hidden_editor_content_without_override_count: hiddenEditorContent.missing_editor_override_count,
-    css_hidden_editor_content: hiddenEditorContent.hidden_rules,
-    css_editor_reveal_overrides: hiddenEditorContent.editor_override_rules,
-    source_css_file_count: sourceCssFiles.length,
-    source_css_structural_selector_count: structuralSelectorDrift.source_structural_selector_count,
-    source_css_missing_structural_selector_count: structuralSelectorDrift.missing_structural_selector_count,
-    source_css_missing_structural_selectors: structuralSelectorDrift.missing_structural_selectors,
-    generated_theme_ux_quality_pass: reasons.length === 0,
-    generated_theme_ux_quality_failure_count: reasons.length,
-    generated_theme_ux_quality_failure_reasons: reasons,
-    remaining_manual_gates: [
-      'site_editor_canvas_above_fold_visible_text',
-      'footer_utility_links_not_responsive_nav_overlay',
-      'admin_bar_not_obscured_by_fixed_or_sticky_chrome',
-    ],
-    artifact: artifactPath,
-  };
-
-  await writeFile(artifactPath, JSON.stringify(diagnostics, null, 2));
-  return diagnostics;
-}
-
-function uniqueSorted(values) {
-  return [...new Set(values.filter(Boolean))].sort((a, b) => a.localeCompare(b));
-}
-
-function uniqueInOrder(values) {
-  const seen = new Set();
-  const ordered = [];
-  for (const value of values) {
-    const normalized = typeof value === 'string' ? value.trim() : '';
-    const key = normalized.toLowerCase();
-    if (!normalized || seen.has(key)) {
-      continue;
-    }
-    seen.add(key);
-    ordered.push(normalized);
-  }
-  return ordered;
-}
-
-function countRegex(value, pattern) {
-  return typeof value === 'string' ? (value.match(pattern) || []).length : 0;
-}
-
-function normalizeFontFamily(value) {
-  return value
-    .replace(/\\["']/g, '')
-    .replace(/["']/g, '')
-    .trim()
-    .replace(/\s+/g, ' ');
-}
-
-function extractFontFamilies(text) {
-  const families = [];
-  for (const match of text.matchAll(/fonts\.googleapis\.com\/css2?[^"')\s]+/gi)) {
-    try {
-      const url = new URL(match[0].startsWith('http') ? match[0] : `https://${match[0]}`);
-      for (const family of url.searchParams.getAll('family')) {
-        families.push(normalizeFontFamily(family.split(':')[0].replaceAll('+', ' ')));
-      }
-    } catch {
-      // Keep parsing local font-family declarations even if an import URL is malformed.
-    }
-  }
-
-  for (const match of text.matchAll(/font-family\s*:\s*([^;}]+)/gi)) {
-    const declaration = match[1] || '';
-    for (const family of declaration.split(',')) {
-      const normalized = normalizeFontFamily(family);
-      if (!/^(sans-serif|serif|monospace|system-ui|inherit|initial|unset)$/i.test(normalized)) {
-        families.push(normalized);
-      }
-    }
-  }
-
-  return uniqueSorted(families);
-}
-
-function firstFontFamilyFromDeclaration(declaration) {
-  for (const family of String(declaration || '').split(',')) {
-    const normalized = normalizeFontFamily(family);
-    if (!/^(sans-serif|serif|monospace|system-ui|ui-sans-serif|ui-serif|inherit|initial|unset)$/i.test(normalized)) {
-      return normalized;
-    }
-  }
-  return '';
-}
-
-function extractFontFamilyDeclarations(text, selectorPattern) {
-  const declarations = [];
-  const pattern = new RegExp(`${selectorPattern}[^{}]*\\{[^{}]*?font-family\\s*:\\s*([^;}]+)`, 'gi');
-  for (const match of text.matchAll(pattern)) {
-    declarations.push(firstFontFamilyFromDeclaration(match[1]));
-  }
-  return uniqueInOrder(declarations);
-}
-
-function extractTypePairing(text, fontFamilies) {
-  const displayFonts = extractFontFamilyDeclarations(
-    text,
-    '(?:h1|h2|h3|\\.display|\\.headline|\\.hero-title|\\.section-title|\\.title)'
-  );
-  const bodyFonts = extractFontFamilyDeclarations(text, '(?:body|html|:root|\\.site|main|p)');
-  const orderedFonts = uniqueInOrder(fontFamilies);
-  const primaryFont = bodyFonts[0] || orderedFonts[0] || '';
-  const displayFont =
-    displayFonts[0] || orderedFonts.find((font) => font.toLowerCase() !== primaryFont.toLowerCase()) || primaryFont;
-
-  return {
-    primary_font_family: primaryFont,
-    display_font_family: displayFont,
-    type_pairing_signature: [displayFont, primaryFont].filter(Boolean).join(' / '),
-  };
-}
-
-function htmlAttributeValue(attributes, name) {
-  const pattern = new RegExp(`\\b${name}\\s*=\\s*(["'])(.*?)\\1`, 'i');
-  return attributes.match(pattern)?.[2] || '';
-}
-
-function extractSections(html) {
-  const sections = [];
-  for (const match of String(html || '').matchAll(/<section\b([^>]*)>([\s\S]*?)<\/section>/gi)) {
-    const attributes = match[1] || '';
-    const content = match[2] || '';
-    sections.push({
-      attributes,
-      className: htmlAttributeValue(attributes, 'class'),
-      style: htmlAttributeValue(attributes, 'style'),
-      content,
-    });
-  }
-  return sections;
-}
-
-export function extractDesignPatternFingerprint({ sourceHtml = '', cssText = '', fontFamilies = [] } = {}) {
-  const text = [sourceHtml, cssText].join('\n');
-  const sections = extractSections(sourceHtml);
-  const heroGridBackgroundCount = countRegex(
-    text,
-    /(?:hero[-_\s]*(?:grid|mesh|background)|(?:grid|mesh)[-_\s]*hero|hero[\s\S]{0,240}(?:radial-gradient|linear-gradient|background-image|background(?:-size)?\s*:|grid-template|mesh)|(?:background|bg)[-_\s]*mesh)/gi
-  );
-  const panelSections = sections.filter((section) =>
-    /\b(?:panel|section-panel|content-panel|full-bleed|full-width|feature-section|pricing-section|testimonial-section|dark-section|light-section)\b/i.test(
-      `${section.className} ${section.style}`
-    )
-  );
-  const fullWidthSections = sections.filter((section) =>
-    /\b(?:full-bleed|full-width|w-full|wide|alignfull|panel)\b/i.test(`${section.className} ${section.style}`) ||
-    /(?:width\s*:\s*100(?:vw|%)|margin-inline\s*:\s*calc\(|left\s*:\s*50%|right\s*:\s*50%)/i.test(section.style)
-  );
-  const eyebrowLabelCount = countRegex(
-    sourceHtml,
-    /class\s*=\s*["'][^"']*\b(?:eyebrow|kicker|overline|label|section-label|pretitle|subheading-label)\b[^"']*["']/gi
-  );
-  const sectionsWithEyebrowTitle = sections.filter((section) =>
-    /class\s*=\s*["'][^"']*\b(?:eyebrow|kicker|overline|label|section-label|pretitle|subheading-label)\b[^"']*["'][\s\S]{0,700}<h[1-3]\b/i.test(
-      section.content
-    )
-  );
-  const typePairing = extractTypePairing(cssText || text, fontFamilies);
-  const patternTokens = [];
-
-  if (heroGridBackgroundCount > 0) {
-    patternTokens.push('hero-grid-background');
-  }
-  if (fullWidthSections.length >= 3) {
-    patternTokens.push('stacked-full-width-panels');
-  } else if (panelSections.length > 0) {
-    patternTokens.push('panel-sections');
-  }
-  if (eyebrowLabelCount > 0) {
-    patternTokens.push('eyebrow-title-labels');
-  }
-  if (typePairing.type_pairing_signature) {
-    patternTokens.push(`type:${typePairing.type_pairing_signature.toLowerCase()}`);
-  }
-
-  return {
-    hero_grid_background_count: heroGridBackgroundCount,
-    hero_grid_background_present: heroGridBackgroundCount > 0,
-    stacked_full_width_section_count: fullWidthSections.length,
-    panel_section_count: panelSections.length,
-    eyebrow_label_count: eyebrowLabelCount,
-    sections_with_eyebrow_title_count: sectionsWithEyebrowTitle.length,
-    font_family_count: fontFamilies.length,
-    ...typePairing,
-    repetition_signature: patternTokens.join('|'),
-  };
-}
-
-function extractColors(text) {
-  const hexColors = [...text.matchAll(/#[0-9a-f]{3,8}\b/gi)].map((match) => match[0].toLowerCase());
-  const functionalColors = [...text.matchAll(/\b(?:rgb|rgba|hsl|hsla)\([^)]*\)/gi)].map((match) =>
-    match[0].toLowerCase().replace(/\s+/g, '')
-  );
-  return uniqueSorted([...hexColors, ...functionalColors]);
-}
-
-function includesAny(text, patterns) {
-  return patterns.some((pattern) => pattern.test(text));
-}
-
-function extractMotifs(text) {
-  const motifs = [];
-  const checks = {
-    bento_grid: [/\bbento\b/i],
-    cards_grid: [/\bcard(s)?\b/i, /grid-template-columns/i],
-    code_preview: [/code-window/i, /code-preview/i, /<pre\b/i, /<code\b/i],
-    dashboard_mockup: [/dashboard/i, /metric-card/i, /analytics/i],
-    glow_overlay: [/\bglow\b/i, /blur\(/i, /radial-gradient\(/i],
-    marquee: [/\bmarquee\b/i, /ticker/i],
-    pricing: [/\bpricing\b/i, /\bplans?\b/i],
-    social_proof: [/testimonial/i, /customer/i, /trusted by/i],
-    split_hero: [/split-hero/i, /hero-grid/i],
-    terminal_window: [/terminal/i, /traffic-light/i, /window-chrome/i, /code-window/i],
-  };
-
-  for (const [motif, patterns] of Object.entries(checks)) {
-    if (includesAny(text, patterns)) {
-      motifs.push(motif);
-    }
-  }
-
-  return motifs.sort();
-}
-
-function extractPaletteLabels(text, colors) {
-  const labels = [];
-  const lower = text.toLowerCase();
-  const colorText = colors.join(' ');
-  if (/purple|violet|indigo|#6|#7|#8|#9|#a/i.test(`${lower} ${colorText}`) && /lime|chartreuse|#bef|#a3e|#ccff|#d9f99d/i.test(`${lower} ${colorText}`)) {
-    labels.push('purple_lime');
-  }
-  if (/orange|amber|coral|#f59|#fb7|#ff8/i.test(`${lower} ${colorText}`)) {
-    labels.push('warm_orange');
-  }
-  if (/cyan|teal|aqua|#06b6|#14b8|#22d3/i.test(`${lower} ${colorText}`)) {
-    labels.push('cyan_teal');
-  }
-  if (/black|charcoal|slate|#0[0-9a-f]{2,6}|#111|#18181b/i.test(`${lower} ${colorText}`)) {
-    labels.push('dark_base');
-  }
-  return labels.sort();
-}
-
-async function collectDesignFingerprint(sitePath) {
-  const { themeRoot, themeSlug } = await collectLatestGeneratedTheme(sitePath);
-  const sourceHtml = await readIfExists(path.join(sitePath, 'tmp/static-site/index.html'));
-  const cssFiles = themeRoot ? await collectCssFiles(themeRoot) : [];
-  const cssParts = [];
-  for (const file of cssFiles) {
-    cssParts.push(await readIfExists(file));
-  }
-
-  const text = [sourceHtml, ...cssParts].join('\n');
-  const lower = text.toLowerCase();
-  const fontFamilies = extractFontFamilies(text);
-  const colors = extractColors(text);
-  const cssVariables = uniqueSorted([...text.matchAll(/--([a-z0-9-]+)\s*:/gi)].map((match) => match[1].toLowerCase()));
-  const motifs = extractMotifs(text);
-  const paletteLabels = extractPaletteLabels(lower, colors);
-  const patternFingerprint = extractDesignPatternFingerprint({
-    sourceHtml,
-    cssText: cssParts.join('\n'),
-    fontFamilies,
+  return compareSemanticFidelityImpl(importReport, artifactDir, sitePath, {
+    studioPath: STUDIO_PATH,
+    viewport: VISUAL_VIEWPORT,
   });
-
-  return {
-    theme_slug: themeSlug,
-    source_html_present: sourceHtml ? true : false,
-    css_file_count: cssFiles.length,
-    font_families: fontFamilies,
-    dominant_font_family: fontFamilies[0] || '',
-    color_values: colors,
-    css_variables: cssVariables,
-    motifs,
-    palette_labels: paletteLabels,
-    gradient_count: countRegex(text, /(?:linear|radial|conic)-gradient\(/gi),
-    animation_count: countRegex(text, /@keyframes\b|\banimation(?:-[a-z]+)?\s*:/gi),
-    transition_count: countRegex(text, /\btransition(?:-[a-z]+)?\s*:/gi),
-    dark_theme: /#0[0-9a-f]{2,6}|#111|#18181b|#020617|background(?:-color)?\s*:\s*(?:black|rgb\(0[,\s]+0[,\s]+0\))/i.test(text),
-    patterns: patternFingerprint,
-  };
 }
 
 function designFingerprintMetrics(fingerprint) {
@@ -2998,82 +780,6 @@ function validationMetrics(result) {
   return { validateCallCount, validateErrorCount, validatedAllCount };
 }
 
-function countMatches(value, pattern) {
-  return typeof value === 'string' ? (value.match(pattern) || []).length : 0;
-}
-
-export function agentAuthoredBlockMetrics(result) {
-  const toolCalls = Array.isArray(result?.toolCalls) ? result.toolCalls : [];
-  const writeCalls = toolCalls.filter((item) => item && item.name === 'Write');
-  let agentAuthoredWpHtmlOpeners = 0;
-  let agentAuthoredWpBlockComments = 0;
-  let agentAuthoredWpHtmlWriteCalls = 0;
-
-  for (const call of writeCalls) {
-    const content = typeof call?.input?.content === 'string' ? call.input.content : '';
-    const wpHtmlOpeners = countMatches(content, /<!--\s+wp:html\b/g);
-    agentAuthoredWpHtmlOpeners += wpHtmlOpeners;
-    agentAuthoredWpBlockComments += countMatches(content, /<!--\s+\/?wp:/g);
-    if (wpHtmlOpeners > 0) {
-      agentAuthoredWpHtmlWriteCalls++;
-    }
-  }
-
-  return {
-    agent_authored_wp_html_openers: agentAuthoredWpHtmlOpeners,
-    agent_authored_wp_html_write_calls: agentAuthoredWpHtmlWriteCalls,
-    agent_authored_wp_block_comments: agentAuthoredWpBlockComments,
-  };
-}
-
-export function nativeBlockQualityMetrics(quality, authoredBlocks, editorValidation, importReport) {
-  const reasons = [];
-  const bfbFallbackCount = metric(quality?.bfb_fallback_count);
-  const coreHtmlWithoutBfbFallback = metric(quality?.core_html_without_bfb_fallback);
-  const importerQuality = importReport?.report?.quality || {};
-  const importerCoreHtmlBlocks = metric(importerQuality.core_html_block_count);
-  const importerInvalidBlocks = metric(importerQuality.invalid_block_count);
-  const agentAuthoredWpHtmlOpeners = metric(authoredBlocks.agent_authored_wp_html_openers);
-  const invalidEditorBlocks = metric(editorValidation?.invalid_blocks);
-  const targetPagesSeen = metric(quality?.target_pages_seen);
-  const targetPostsWithBlocks = metric(quality?.target_posts_with_blocks);
-
-  if (targetPagesSeen === 0 || targetPostsWithBlocks === 0) {
-    reasons.push('missing_target_block_page');
-  }
-
-  if (agentAuthoredWpHtmlOpeners > 0) {
-    reasons.push('agent_authored_wp_html');
-  }
-  if (coreHtmlWithoutBfbFallback > 0) {
-    reasons.push('core_html_without_bfb_fallback');
-  }
-  if (bfbFallbackCount > 0) {
-    reasons.push('bfb_fallback');
-  }
-  if (importerCoreHtmlBlocks > 0) {
-    reasons.push('importer_core_html_blocks');
-  }
-  if (importerInvalidBlocks > 0) {
-    reasons.push('importer_invalid_blocks');
-  }
-  if (importReport?.error) {
-    reasons.push('importer_report_error');
-  }
-  if (invalidEditorBlocks > 0) {
-    reasons.push('editor_invalid_blocks');
-  }
-  if (editorValidation?.error) {
-    reasons.push('editor_validation_error');
-  }
-
-  return {
-    native_block_quality_pass: reasons.length === 0,
-    native_block_quality_failure_count: reasons.length,
-    native_block_quality_failure_reasons: reasons,
-  };
-}
-
 function metric(value) {
   const number = Number(value ?? 0);
   return Number.isFinite(number) ? number : 0;
@@ -3111,25 +817,6 @@ function optionalArtifactPath(name, value) {
   return typeof value === 'string' && value.length > 0 ? { [name]: value } : {};
 }
 
-export function semanticTargetMetric(semanticComparison, key) {
-  return (semanticComparison?.diagnostics?.targets || semanticComparison?.targets || []).reduce(
-    (sum, target) => sum + metric(target?.[key]),
-    0
-  );
-}
-
-export function semanticMismatchFailureDetails(semanticComparison) {
-  const mismatches = semanticComparison?.diagnostics?.mismatches || semanticComparison?.mismatches || [];
-  return mismatches.map((mismatch) => {
-    const concept = mismatch.concept || mismatch.type || 'unknown';
-    const sourceCount =
-      mismatch.source && Object.hasOwn(mismatch.source, 'count') ? ` source=${mismatch.source.count}` : '';
-    const generatedCount =
-      mismatch.generated && Object.hasOwn(mismatch.generated, 'count') ? ` generated=${mismatch.generated.count}` : '';
-    const reason = mismatch.reason ? ` reason=${mismatch.reason}` : '';
-    return `semantic mismatch: ${concept}${sourceCount}${generatedCount}${reason}`;
-  });
-}
 
 export function agentSuccessGate(result, semanticComparison) {
   const semanticMismatchCount = metric(semanticComparison?.mismatch_count);
@@ -3167,14 +854,16 @@ export default async function studioAgentSiteBuildBench() {
   const selectedPromptFile = String(await promptTemplatePath());
   const systemPrompt = await systemPromptFingerprint();
   const prompt = await siteBuildPrompt(sitePath);
+  const model = evalModel();
   const agentStarted = Date.now();
   const { result, resultFile, exitCode, stderr } = await runEval(prompt, {
     maxTurns: 40,
     timeoutMs: 420000,
+    ...(model ? { model } : {}),
   });
   const agentElapsedMs = Date.now() - agentStarted;
   const qualityProbeStarted = Date.now();
-  const quality = await probeQuality(sitePath);
+  const quality = await probeQuality(sitePath, { runCli });
   const qualityProbeMs = Date.now() - qualityProbeStarted;
   const status = await siteStatus(sitePath);
   assertNamespacedPort(status, runtime);
@@ -3215,6 +904,7 @@ export default async function studioAgentSiteBuildBench() {
         prompt_variant: selectedPromptVariant,
         prompt_file: selectedPromptFile,
         prompt_category: PROMPT_CATEGORY,
+        model,
         ...systemPrompt,
         prompt,
         sitePath,
@@ -3390,6 +1080,7 @@ export default async function studioAgentSiteBuildBench() {
       prompt_variant: selectedPromptVariant,
       prompt_file: selectedPromptFile,
       prompt_category: PROMPT_CATEGORY,
+      model: model || 'default',
       design_primary_font_family: designFingerprint.patterns?.primary_font_family || '',
       design_display_font_family: designFingerprint.patterns?.display_font_family || '',
       design_type_pairing_signature: designFingerprint.patterns?.type_pairing_signature || '',
